% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityCreationDTO.r
\docType{data}
\name{ActivityCreationDTO}
\alias{ActivityCreationDTO}
\title{ActivityCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActivityCreationDTO
}
\description{
ActivityCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdf_type}}{}

\item{\code{uri}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{timezone}}{}

\item{\code{settings}}{}
}}

\keyword{datasets}
