% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityGetDTO.r
\docType{data}
\name{ActivityGetDTO}
\alias{ActivityGetDTO}
\title{ActivityGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActivityGetDTO
}
\description{
ActivityGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdf_type}}{}

\item{\code{uri}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{settings}}{}
}}

\keyword{datasets}
