% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCreationDTO.r
\docType{data}
\name{AnnotationCreationDTO}
\alias{AnnotationCreationDTO}
\title{AnnotationCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationCreationDTO
}
\description{
AnnotationCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{motivation}}{}
}}

\keyword{datasets}
