% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGetDTO.r
\docType{data}
\name{AnnotationGetDTO}
\alias{AnnotationGetDTO}
\title{AnnotationGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationGetDTO
}
\description{
AnnotationGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{motivation}}{}

\item{\code{created}}{}

\item{\code{author}}{}
}}

\keyword{datasets}
