% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationsApi.r
\docType{data}
\name{AnnotationsApi}
\alias{AnnotationsApi}
\title{Annotations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationsApi
}
\description{
opensilexClientToolsR.Annotations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count_annotations Count annotations


create_annotation Create an annotation


delete_annotation Delete an annotation


get_annotation Get an annotation


search_annotations Search annotations


search_motivations Search motivations


update_annotation Update an annotation

}
}

\keyword{datasets}
