% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationUpdateDTO.r
\docType{data}
\name{AnnotationUpdateDTO}
\alias{AnnotationUpdateDTO}
\title{AnnotationUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationUpdateDTO
}
\description{
AnnotationUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{motivation}}{}
}}

\keyword{datasets}
