% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiGitCommitDTO.r
\docType{data}
\name{ApiGitCommitDTO}
\alias{ApiGitCommitDTO}
\title{ApiGitCommitDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiGitCommitDTO
}
\description{
ApiGitCommitDTO Class
}
\section{Fields}{

\describe{
\item{\code{commit_id}}{}

\item{\code{commit_message}}{}
}}

\keyword{datasets}
