% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiLicenseInfoDTO.r
\docType{data}
\name{ApiLicenseInfoDTO}
\alias{ApiLicenseInfoDTO}
\title{ApiLicenseInfoDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiLicenseInfoDTO
}
\description{
ApiLicenseInfoDTO Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{url}}{}
}}

\keyword{datasets}
