% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiModulesInfo.r
\docType{data}
\name{ApiModulesInfo}
\alias{ApiModulesInfo}
\title{ApiModulesInfo Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiModulesInfo
}
\description{
ApiModulesInfo Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{version}}{}
}}

\keyword{datasets}
