% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaApi.r
\name{AreaApi}
\alias{AreaApi}
\title{Area operations}
\description{
opensilexClientToolsR.Area
}
\section{Methods}{

\describe{

create_area Add an area


delete_area Delete an area


get_by_uri Get an area


search_intersects Get area whose geometry corresponds to the Intersections


update_area Update an area

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AreaApi$new()}}
\item \href{#method-create_area}{\code{AreaApi$create_area()}}
\item \href{#method-delete_area}{\code{AreaApi$delete_area()}}
\item \href{#method-get_by_uri}{\code{AreaApi$get_by_uri()}}
\item \href{#method-search_intersects}{\code{AreaApi$search_intersects()}}
\item \href{#method-update_area}{\code{AreaApi$update_area()}}
\item \href{#method-clone}{\code{AreaApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_area"></a>}}
\if{latex}{\out{\hypertarget{method-create_area}{}}}
\subsection{Method \code{create_area()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$create_area(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_area"></a>}}
\if{latex}{\out{\hypertarget{method-delete_area}{}}}
\subsection{Method \code{delete_area()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$delete_area(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_by_uri}{}}}
\subsection{Method \code{get_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$get_by_uri(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_intersects"></a>}}
\if{latex}{\out{\hypertarget{method-search_intersects}{}}}
\subsection{Method \code{search_intersects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$search_intersects(body, start, end, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_area"></a>}}
\if{latex}{\out{\hypertarget{method-update_area}{}}}
\subsection{Method \code{update_area()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$update_area(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AreaApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
