% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaCreationDTO.r
\docType{data}
\name{AreaCreationDTO}
\alias{AreaCreationDTO}
\title{AreaCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AreaCreationDTO
}
\description{
AreaCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{geometry}}{}
}}

\keyword{datasets}
