% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaGetDTO.r
\docType{data}
\name{AreaGetDTO}
\alias{AreaGetDTO}
\title{AreaGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AreaGetDTO
}
\description{
AreaGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{author}}{}

\item{\code{geometry}}{}
}}

\keyword{datasets}
