% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaUpdateDTO.r
\docType{data}
\name{AreaUpdateDTO}
\alias{AreaUpdateDTO}
\title{AreaUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AreaUpdateDTO
}
\description{
AreaUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{geometry}}{}
}}

\keyword{datasets}
