% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationApi.r
\name{AuthenticationApi}
\alias{AuthenticationApi}
\title{Authentication operations}
\description{
opensilexClientToolsR.Authentication
}
\section{Methods}{

\describe{

authenticate Authenticate a user and return an access token


authenticate_open_id Authenticate a user and return an access token


forgot_password Send an e-mail confirmation


get_credentials_groups Get list of existing credentials indexed by Swagger @API concepts in the application


logout Logout by discarding a user token


renew_password Update user password


renew_token Send back a new token if the provided one is still valid

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AuthenticationApi$new()}}
\item \href{#method-authenticate}{\code{AuthenticationApi$authenticate()}}
\item \href{#method-authenticate_open_id}{\code{AuthenticationApi$authenticate_open_id()}}
\item \href{#method-forgot_password}{\code{AuthenticationApi$forgot_password()}}
\item \href{#method-get_credentials_groups}{\code{AuthenticationApi$get_credentials_groups()}}
\item \href{#method-logout}{\code{AuthenticationApi$logout()}}
\item \href{#method-renew_password}{\code{AuthenticationApi$renew_password()}}
\item \href{#method-renew_token}{\code{AuthenticationApi$renew_token()}}
\item \href{#method-clone}{\code{AuthenticationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate"></a>}}
\if{latex}{\out{\hypertarget{method-authenticate}{}}}
\subsection{Method \code{authenticate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$authenticate(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate_open_id"></a>}}
\if{latex}{\out{\hypertarget{method-authenticate_open_id}{}}}
\subsection{Method \code{authenticate_open_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$authenticate_open_id(code, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forgot_password"></a>}}
\if{latex}{\out{\hypertarget{method-forgot_password}{}}}
\subsection{Method \code{forgot_password()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$forgot_password(identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_credentials_groups"></a>}}
\if{latex}{\out{\hypertarget{method-get_credentials_groups}{}}}
\subsection{Method \code{get_credentials_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$get_credentials_groups(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logout"></a>}}
\if{latex}{\out{\hypertarget{method-logout}{}}}
\subsection{Method \code{logout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$logout(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renew_password"></a>}}
\if{latex}{\out{\hypertarget{method-renew_password}{}}}
\subsection{Method \code{renew_password()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$renew_password(renew_token, check_only, password, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renew_token"></a>}}
\if{latex}{\out{\hypertarget{method-renew_token}{}}}
\subsection{Method \code{renew_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$renew_token(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
