% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationDTO.r
\docType{data}
\name{AuthenticationDTO}
\alias{AuthenticationDTO}
\title{AuthenticationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AuthenticationDTO
}
\description{
AuthenticationDTO Class
}
\section{Fields}{

\describe{
\item{\code{identifier}}{}

\item{\code{password}}{}
}}

\keyword{datasets}
