% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAPIApi.r
\name{BRAPIApi}
\alias{BRAPIApi}
\title{BRAPI operations}
\description{
opensilexClientToolsR.BRAPI
}
\section{Methods}{

\describe{

get_calls Check the available BrAPI calls


get_germplasm_by_search Submit a search request for germplasm


get_observation_units List all the observation units measured in the study.


get_observation_variables List all the observation variables measured in the study.


get_observations Get the observations associated to a specific study


get_studies Retrieve studies information


get_studies_search Retrieve studies information


get_study_details Retrieve study details


get_variable_details Retrieve variable details by id


get_variables_list Call to retrieve a list of observationVariables available in the system

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BRAPIApi$new()}}
\item \href{#method-get_calls}{\code{BRAPIApi$get_calls()}}
\item \href{#method-get_germplasm_by_search}{\code{BRAPIApi$get_germplasm_by_search()}}
\item \href{#method-get_observation_units}{\code{BRAPIApi$get_observation_units()}}
\item \href{#method-get_observation_variables}{\code{BRAPIApi$get_observation_variables()}}
\item \href{#method-get_observations}{\code{BRAPIApi$get_observations()}}
\item \href{#method-get_studies}{\code{BRAPIApi$get_studies()}}
\item \href{#method-get_studies_search}{\code{BRAPIApi$get_studies_search()}}
\item \href{#method-get_study_details}{\code{BRAPIApi$get_study_details()}}
\item \href{#method-get_variable_details}{\code{BRAPIApi$get_variable_details()}}
\item \href{#method-get_variables_list}{\code{BRAPIApi$get_variables_list()}}
\item \href{#method-clone}{\code{BRAPIApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_calls"></a>}}
\if{latex}{\out{\hypertarget{method-get_calls}{}}}
\subsection{Method \code{get_calls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_calls(page, page_size, data_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_germplasm_by_search"></a>}}
\if{latex}{\out{\hypertarget{method-get_germplasm_by_search}{}}}
\subsection{Method \code{get_germplasm_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_germplasm_by_search(
  germplasm_db_id,
  germplasm_pui,
  germplasm_name,
  common_crop_name,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observation_units"></a>}}
\if{latex}{\out{\hypertarget{method-get_observation_units}{}}}
\subsection{Method \code{get_observation_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_observation_units(
  study_db_id,
  observation_level,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observation_variables"></a>}}
\if{latex}{\out{\hypertarget{method-get_observation_variables}{}}}
\subsection{Method \code{get_observation_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_observation_variables(study_db_id, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observations"></a>}}
\if{latex}{\out{\hypertarget{method-get_observations}{}}}
\subsection{Method \code{get_observations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_observations(
  study_db_id,
  observation_variable_db_ids,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_studies"></a>}}
\if{latex}{\out{\hypertarget{method-get_studies}{}}}
\subsection{Method \code{get_studies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_studies(
  study_db_id,
  active,
  sort_by,
  sort_order,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_studies_search"></a>}}
\if{latex}{\out{\hypertarget{method-get_studies_search}{}}}
\subsection{Method \code{get_studies_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_studies_search(
  study_db_id,
  active,
  sort_by,
  sort_order,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_study_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_study_details}{}}}
\subsection{Method \code{get_study_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_study_details(study_db_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_variable_details}{}}}
\subsection{Method \code{get_variable_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_variable_details(observation_variable_db_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_variables_list}{}}}
\subsection{Method \code{get_variables_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$get_variables_list(observation_variable_db_id, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BRAPIApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
