% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacteristicDetailsDTO.r
\docType{data}
\name{CharacteristicDetailsDTO}
\alias{CharacteristicDetailsDTO}
\title{CharacteristicDetailsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CharacteristicDetailsDTO
}
\description{
CharacteristicDetailsDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
