% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacteristicGetDTO.r
\docType{data}
\name{CharacteristicGetDTO}
\alias{CharacteristicGetDTO}
\title{CharacteristicGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CharacteristicGetDTO
}
\description{
CharacteristicGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
