% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcernedItemPositionCreationDTO.r
\docType{data}
\name{ConcernedItemPositionCreationDTO}
\alias{ConcernedItemPositionCreationDTO}
\title{ConcernedItemPositionCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConcernedItemPositionCreationDTO
}
\description{
ConcernedItemPositionCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{target}}{}

\item{\code{position}}{}
}}

\keyword{datasets}
