% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcernedItemPositionGetDTO.r
\docType{data}
\name{ConcernedItemPositionGetDTO}
\alias{ConcernedItemPositionGetDTO}
\title{ConcernedItemPositionGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConcernedItemPositionGetDTO
}
\description{
ConcernedItemPositionGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{target}}{}

\item{\code{position}}{}
}}

\keyword{datasets}
