% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiClient.r
\name{connectToOpenSILEX}
\alias{connectToOpenSILEX}
\title{connectToOpenSILEX}
\usage{
connectToOpenSILEX(
  identifier = NULL,
  password = NULL,
  url = NULL,
  reconnection = TRUE
)
}
\arguments{
\item{identifier}{login of the user to create the token}

\item{password}{password of the user to create the token}

\item{url}{character, if apiID is private add the url of the chosen API, containing the IP,
the full url with the protocol. e.g. 'http://www.opensilex.org/openSilexAPI/rest'}

\item{reconnection}{to force the client reconnection}
}
\description{
load name space and connexion parameters of the webservice.
Execute only once at the beginning of the requests.
In the case of a WebService change of address or a renaming of services, please edit this list.
and execute the function.
Demonstration instances :
\describe{
connectToOpenSILEX(identifier="guest@opensilex.org",password="guest",
url = "http://www.opensilex.org/openSilexAPI/rest")
}
}
