% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialDTO.r
\docType{data}
\name{CredentialDTO}
\alias{CredentialDTO}
\title{CredentialDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialDTO
}
\description{
CredentialDTO Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
