% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsGroupDTO.r
\docType{data}
\name{CredentialsGroupDTO}
\alias{CredentialsGroupDTO}
\title{CredentialsGroupDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsGroupDTO
}
\description{
CredentialsGroupDTO Class
}
\section{Fields}{

\describe{
\item{\code{group_id}}{}

\item{\code{group_key_name}}{}

\item{\code{credentials}}{}
}}

\keyword{datasets}
