% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVCell.r
\docType{data}
\name{CSVCell}
\alias{CSVCell}
\title{CSVCell Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVCell
}
\description{
CSVCell Class
}
\section{Fields}{

\describe{
\item{\code{rowIndex}}{}

\item{\code{colIndex}}{}

\item{\code{header}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
