% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVDatatypeError.r
\docType{data}
\name{CSVDatatypeError}
\alias{CSVDatatypeError}
\title{CSVDatatypeError Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVDatatypeError
}
\description{
CSVDatatypeError Class
}
\section{Fields}{

\describe{
\item{\code{rowIndex}}{}

\item{\code{colIndex}}{}

\item{\code{header}}{}

\item{\code{value}}{}

\item{\code{datatype}}{}
}}

\keyword{datasets}
