% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVDuplicateURIError.r
\docType{data}
\name{CSVDuplicateURIError}
\alias{CSVDuplicateURIError}
\title{CSVDuplicateURIError Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVDuplicateURIError
}
\description{
CSVDuplicateURIError Class
}
\section{Fields}{

\describe{
\item{\code{rowIndex}}{}

\item{\code{colIndex}}{}

\item{\code{header}}{}

\item{\code{value}}{}

\item{\code{previousRow}}{}
}}

\keyword{datasets}
