% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVURINotFoundError.r
\name{CSVURINotFoundError}
\alias{CSVURINotFoundError}
\title{CSVURINotFoundError Class}
\description{
CSVURINotFoundError Class

CSVURINotFoundError Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rowIndex}}{}

\item{\code{colIndex}}{}

\item{\code{header}}{}

\item{\code{value}}{}

\item{\code{rdfType}}{}

\item{\code{objectURI}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSVURINotFoundError$new()}}
\item \href{#method-toJSON}{\code{CSVURINotFoundError$toJSON()}}
\item \href{#method-fromJSON}{\code{CSVURINotFoundError$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{CSVURINotFoundError$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{CSVURINotFoundError$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CSVURINotFoundError$fromJSONString()}}
\item \href{#method-clone}{\code{CSVURINotFoundError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$new(rowIndex, colIndex, header, value, rdfType, objectURI)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$fromJSON(CSVURINotFoundErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$fromJSONObject(CSVURINotFoundErrorObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$fromJSONString(CSVURINotFoundErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVURINotFoundError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
