% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVValidationDTO.r
\docType{data}
\name{CSVValidationDTO}
\alias{CSVValidationDTO}
\title{CSVValidationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVValidationDTO
}
\description{
CSVValidationDTO Class
}
\section{Fields}{

\describe{
\item{\code{errors}}{}

\item{\code{validation_token}}{}

\item{\code{nb_lines_imported}}{}
}}

\keyword{datasets}
