% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVValidationModel.r
\name{CSVValidationModel}
\alias{CSVValidationModel}
\title{CSVValidationModel Class}
\description{
CSVValidationModel Class

CSVValidationModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{missingHeaders}}{}

\item{\code{emptyHeaders}}{}

\item{\code{invalidHeaderURIs}}{}

\item{\code{datatypeErrors}}{}

\item{\code{uriNotFoundErrors}}{}

\item{\code{invalidURIErrors}}{}

\item{\code{missingRequiredValueErrors}}{}

\item{\code{invalidValueErrors}}{}

\item{\code{alreadyExistingURIErrors}}{}

\item{\code{duplicateURIErrors}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSVValidationModel$new()}}
\item \href{#method-toJSON}{\code{CSVValidationModel$toJSON()}}
\item \href{#method-fromJSON}{\code{CSVValidationModel$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{CSVValidationModel$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{CSVValidationModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CSVValidationModel$fromJSONString()}}
\item \href{#method-clone}{\code{CSVValidationModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$new(
  missingHeaders,
  emptyHeaders,
  invalidHeaderURIs,
  datatypeErrors,
  uriNotFoundErrors,
  invalidURIErrors,
  missingRequiredValueErrors,
  invalidValueErrors,
  alreadyExistingURIErrors,
  duplicateURIErrors
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$fromJSON(CSVValidationModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$fromJSONObject(CSVValidationModelObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$fromJSONString(CSVValidationModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVValidationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
