% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataApi.r
\docType{data}
\name{DataApi}
\alias{DataApi}
\title{Data operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataApi
}
\description{
opensilexClientToolsR.Data
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

add_list_data Add data


create_provenance Add a provenance


delete_data Delete data


delete_data_on_search Delete data on criteria


delete_provenance Delete a provenance that doesn&#39;t describe data


get_data Get data


get_data_file Get a data file


get_data_file_description Get a data file description


get_data_file_descriptions_by_search Search data files


get_pictures_thumbnails Get a picture thumbnail


get_provenance Get a provenance


get_provenances_by_ur_is Get a list of provenances by their URIs


post_data_file Add a data file


post_data_file_paths Describe datafiles and give their relative paths in the configured storage system. In the case of already stored datafiles.


search_data_list Search data


search_provenance Get provenances


update Update data


update1 Update a provenance


update_confidence Update confidence index

}
}

\keyword{datasets}
