% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConfidenceDTO.r
\docType{data}
\name{DataConfidenceDTO}
\alias{DataConfidenceDTO}
\title{DataConfidenceDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataConfidenceDTO
}
\description{
DataConfidenceDTO Class
}
\section{Fields}{

\describe{
\item{\code{confidence}}{}
}}

\keyword{datasets}
