% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCSVValidationDTO.r
\docType{data}
\name{DataCSVValidationDTO}
\alias{DataCSVValidationDTO}
\title{DataCSVValidationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataCSVValidationDTO
}
\description{
DataCSVValidationDTO Class
}
\section{Fields}{

\describe{
\item{\code{errors}}{}

\item{\code{dataErrors}}{}

\item{\code{sizeMax}}{}

\item{\code{validation_token}}{}

\item{\code{nb_lines_imported}}{}
}}

\keyword{datasets}
