% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCSVValidationModel.r
\name{DataCSVValidationModel}
\alias{DataCSVValidationModel}
\title{DataCSVValidationModel Class}
\description{
DataCSVValidationModel Class

DataCSVValidationModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{missingHeaders}}{}

\item{\code{emptyHeaders}}{}

\item{\code{invalidHeaderURIs}}{}

\item{\code{datatypeErrors}}{}

\item{\code{uriNotFoundErrors}}{}

\item{\code{invalidURIErrors}}{}

\item{\code{missingRequiredValueErrors}}{}

\item{\code{invalidValueErrors}}{}

\item{\code{alreadyExistingURIErrors}}{}

\item{\code{duplicateURIErrors}}{}

\item{\code{invalidObjectErrors}}{}

\item{\code{invalidDateErrors}}{}

\item{\code{invalidDataTypeErrors}}{}

\item{\code{invalidExperimentErrors}}{}

\item{\code{invalidDeviceErrors}}{}

\item{\code{duplicatedDataErrors}}{}

\item{\code{duplicatedObjectErrors}}{}

\item{\code{duplicatedExperimentErrors}}{}

\item{\code{duplicatedDeviceErrors}}{}

\item{\code{headers}}{}

\item{\code{headersLabels}}{}

\item{\code{nbLinesImported}}{}

\item{\code{nbLinesToImport}}{}

\item{\code{validationStep}}{}

\item{\code{insertionStep}}{}

\item{\code{validCSV}}{}

\item{\code{tooLargeDataset}}{}

\item{\code{errorMessage}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataCSVValidationModel$new()}}
\item \href{#method-toJSON}{\code{DataCSVValidationModel$toJSON()}}
\item \href{#method-fromJSON}{\code{DataCSVValidationModel$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{DataCSVValidationModel$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{DataCSVValidationModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DataCSVValidationModel$fromJSONString()}}
\item \href{#method-clone}{\code{DataCSVValidationModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$new(
  missingHeaders,
  emptyHeaders,
  invalidHeaderURIs,
  datatypeErrors,
  uriNotFoundErrors,
  invalidURIErrors,
  missingRequiredValueErrors,
  invalidValueErrors,
  alreadyExistingURIErrors,
  duplicateURIErrors,
  invalidObjectErrors,
  invalidDateErrors,
  invalidDataTypeErrors,
  invalidExperimentErrors,
  invalidDeviceErrors,
  duplicatedDataErrors,
  duplicatedObjectErrors,
  duplicatedExperimentErrors,
  duplicatedDeviceErrors,
  headers,
  headersLabels,
  nbLinesImported,
  nbLinesToImport,
  validationStep,
  insertionStep,
  validCSV,
  tooLargeDataset,
  errorMessage
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$fromJSON(DataCSVValidationModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$fromJSONObject(DataCSVValidationModelObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$fromJSONString(DataCSVValidationModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCSVValidationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
