% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeviceCreationDTO.r
\docType{data}
\name{DeviceCreationDTO}
\alias{DeviceCreationDTO}
\title{DeviceCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DeviceCreationDTO
}
\description{
DeviceCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{brand}}{}

\item{\code{constructor_model}}{}

\item{\code{serial_number}}{}

\item{\code{person_in_charge}}{}

\item{\code{start_up}}{}

\item{\code{removal}}{}

\item{\code{relations}}{}

\item{\code{description}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
