% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeviceGetDetailsDTO.r
\name{DeviceGetDetailsDTO}
\alias{DeviceGetDetailsDTO}
\title{DeviceGetDetailsDTO Class}
\description{
DeviceGetDetailsDTO Class

DeviceGetDetailsDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{brand}}{}

\item{\code{constructor_model}}{}

\item{\code{serial_number}}{}

\item{\code{person_in_charge}}{}

\item{\code{start_up}}{}

\item{\code{removal}}{}

\item{\code{relations}}{}

\item{\code{description}}{}

\item{\code{metadata}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeviceGetDetailsDTO$new()}}
\item \href{#method-toJSON}{\code{DeviceGetDetailsDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{DeviceGetDetailsDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{DeviceGetDetailsDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{DeviceGetDetailsDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DeviceGetDetailsDTO$fromJSONString()}}
\item \href{#method-clone}{\code{DeviceGetDetailsDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$new(
  uri,
  rdf_type,
  rdf_type_name,
  name,
  brand,
  constructor_model,
  serial_number,
  person_in_charge,
  start_up,
  removal,
  relations,
  description,
  metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$fromJSON(DeviceGetDetailsDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$fromJSONObject(DeviceGetDetailsDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$fromJSONString(DeviceGetDetailsDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeviceGetDetailsDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
