% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevicesApi.r
\docType{data}
\name{DevicesApi}
\alias{DevicesApi}
\title{Devices operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DevicesApi
}
\description{
opensilexClientToolsR.Devices
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_device Create a device


delete_device Delete a device


export_devices export devices


export_list export devices


get_device Get device detail


get_device_data_files_provenances Get provenances of datafiles linked to this device


get_device_data_provenances Get provenances of data that have been measured on this device


get_device_variables Get variables measured by the device


search_device_data Search device data


search_device_datafiles Search device datafiles descriptions


search_devices Search devices


update_device Update a device

}
}

\keyword{datasets}
