% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationLink.r
\docType{data}
\name{DocumentationLink}
\alias{DocumentationLink}
\title{DocumentationLink Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentationLink
}
\description{
DocumentationLink Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{url}}{}
}}

\keyword{datasets}
