% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentsApi.r
\name{DocumentsApi}
\alias{DocumentsApi}
\title{Documents operations}
\description{
opensilexClientToolsR.Documents
}
\section{Methods}{

\describe{

create_document Add a document


delete_document Delete a document


get_document_file Get document


get_document_metadata Get document&#39;s description


search_documents Search documents


update_document Update document&#39;s description

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DocumentsApi$new()}}
\item \href{#method-create_document}{\code{DocumentsApi$create_document()}}
\item \href{#method-delete_document}{\code{DocumentsApi$delete_document()}}
\item \href{#method-get_document_file}{\code{DocumentsApi$get_document_file()}}
\item \href{#method-get_document_metadata}{\code{DocumentsApi$get_document_metadata()}}
\item \href{#method-search_documents}{\code{DocumentsApi$search_documents()}}
\item \href{#method-update_document}{\code{DocumentsApi$update_document()}}
\item \href{#method-clone}{\code{DocumentsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_document"></a>}}
\if{latex}{\out{\hypertarget{method-create_document}{}}}
\subsection{Method \code{create_document()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$create_document(description, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_document"></a>}}
\if{latex}{\out{\hypertarget{method-delete_document}{}}}
\subsection{Method \code{delete_document()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$delete_document(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_document_file"></a>}}
\if{latex}{\out{\hypertarget{method-get_document_file}{}}}
\subsection{Method \code{get_document_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$get_document_file(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_document_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_document_metadata}{}}}
\subsection{Method \code{get_document_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$get_document_metadata(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_documents"></a>}}
\if{latex}{\out{\hypertarget{method-search_documents}{}}}
\subsection{Method \code{search_documents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$search_documents(
  rdf_type,
  title,
  date,
  targets,
  authors,
  keyword,
  multiple,
  deprecated,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_document"></a>}}
\if{latex}{\out{\hypertarget{method-update_document}{}}}
\subsection{Method \code{update_document()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$update_document(description, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
