% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityUpdateDTO.r
\name{EntityUpdateDTO}
\alias{EntityUpdateDTO}
\title{EntityUpdateDTO Class}
\description{
EntityUpdateDTO Class

EntityUpdateDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EntityUpdateDTO$new()}}
\item \href{#method-toJSON}{\code{EntityUpdateDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{EntityUpdateDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{EntityUpdateDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{EntityUpdateDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{EntityUpdateDTO$fromJSONString()}}
\item \href{#method-clone}{\code{EntityUpdateDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$new(
  uri,
  name,
  description,
  exact_match,
  close_match,
  broad_match,
  narrow_match
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$fromJSON(EntityUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$fromJSONObject(EntityUpdateDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$fromJSONString(EntityUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityUpdateDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
