% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorDTO.r
\docType{data}
\name{ErrorDTO}
\alias{ErrorDTO}
\title{ErrorDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ErrorDTO
}
\description{
ErrorDTO Class
}
\section{Fields}{

\describe{
\item{\code{title}}{}

\item{\code{message}}{}
}}

\keyword{datasets}
