% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorResponse.r
\docType{data}
\name{ErrorResponse}
\alias{ErrorResponse}
\title{ErrorResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ErrorResponse
}
\description{
ErrorResponse Class
}
\section{Fields}{

\describe{
\item{\code{metadata}}{}

\item{\code{result}}{}
}}

\keyword{datasets}
