% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventsApi.r
\docType{data}
\name{EventsApi}
\alias{EventsApi}
\title{Events operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventsApi
}
\description{
opensilexClientToolsR.Events
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count_events Count events


create_events Create a list of event


create_moves Create a list of move event


delete_event Delete an event


delete_move_event Delete a move event


get_event Get an event


get_event_details Get an event with all it&#39;s properties


get_move_event Get a move with all it&#39;s properties


import_event_csv Import a CSV file with one move and one target per line


import_move_csv Import a CSV file with one move and one target per line


search_events Search events


update_event Update an event


update_move_event Update a move event


validate_event_csv Check a CSV file with one move and one target per line


validate_move_csv Check a CSV file with one move and one target per line

}
}

\keyword{datasets}
