% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventUpdateDTO.r
\docType{data}
\name{EventUpdateDTO}
\alias{EventUpdateDTO}
\title{EventUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventUpdateDTO
}
\description{
EventUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{relations}}{}
}}

\keyword{datasets}
