% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentCreationDTO.r
\name{ExperimentCreationDTO}
\alias{ExperimentCreationDTO}
\title{ExperimentCreationDTO Class}
\description{
ExperimentCreationDTO Class

ExperimentCreationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{description}}{}

\item{\code{objective}}{}

\item{\code{species}}{}

\item{\code{variables}}{}

\item{\code{organisations}}{}

\item{\code{projects}}{}

\item{\code{scientific_supervisors}}{}

\item{\code{technical_supervisors}}{}

\item{\code{groups}}{}

\item{\code{factors}}{}

\item{\code{is_public}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExperimentCreationDTO$new()}}
\item \href{#method-toJSON}{\code{ExperimentCreationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ExperimentCreationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ExperimentCreationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ExperimentCreationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ExperimentCreationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ExperimentCreationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$new(
  uri,
  name,
  start_date,
  end_date,
  description,
  objective,
  species,
  variables,
  organisations,
  projects,
  scientific_supervisors,
  technical_supervisors,
  groups,
  factors,
  is_public
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$fromJSON(ExperimentCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$fromJSONObject(ExperimentCreationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$fromJSONString(ExperimentCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentCreationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
