% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentGetDTO.r
\docType{data}
\name{ExperimentGetDTO}
\alias{ExperimentGetDTO}
\title{ExperimentGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExperimentGetDTO
}
\description{
ExperimentGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{description}}{}

\item{\code{objective}}{}

\item{\code{species}}{}

\item{\code{factors}}{}

\item{\code{organisations}}{}

\item{\code{facilities}}{}

\item{\code{projects}}{}

\item{\code{scientific_supervisors}}{}

\item{\code{technical_supervisors}}{}

\item{\code{groups}}{}

\item{\code{is_public}}{}
}}

\keyword{datasets}
