% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentGetListDTO.r
\docType{data}
\name{ExperimentGetListDTO}
\alias{ExperimentGetListDTO}
\title{ExperimentGetListDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExperimentGetListDTO
}
\description{
ExperimentGetListDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{description}}{}

\item{\code{objective}}{}

\item{\code{species}}{}

\item{\code{is_public}}{}
}}

\keyword{datasets}
