% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentsApi.r
\docType{data}
\name{ExperimentsApi}
\alias{ExperimentsApi}
\title{Experiments operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExperimentsApi
}
\description{
opensilexClientToolsR.Experiments
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_experiment Add an experiment


delete_experiment Delete an experiment


export_experiment_data_list export experiment data


get_available_facilities Get facilities available for an experiment


get_available_factors Get factors with their levels associated to an experiment


get_available_species Get species present in an experiment


get_experiment Get an experiment


get_experiments_by_ur_is Get experiments URIs


get_used_variables1 Get variables involved in an experiment


import_csv_data1 Import a CSV file for the given experiment URI and scientific object type.


search_experiment_data_list Search data


search_experiment_provenances Get provenances involved in an experiment


search_experiments Search experiments


update_experiment Update an experiment


validate_csv1 Import a CSV file for the given experiment URI and scientific object type.

}
}

\keyword{datasets}
