% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacilityAddressDTO.r
\docType{data}
\name{FacilityAddressDTO}
\alias{FacilityAddressDTO}
\title{FacilityAddressDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FacilityAddressDTO
}
\description{
FacilityAddressDTO Class
}
\section{Fields}{

\describe{
\item{\code{countryName}}{}

\item{\code{locality}}{}

\item{\code{postalCode}}{}

\item{\code{region}}{}

\item{\code{streetAddress}}{}

\item{\code{readableAddress}}{}
}}

\keyword{datasets}
