% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorCategoryGetDTO.r
\docType{data}
\name{FactorCategoryGetDTO}
\alias{FactorCategoryGetDTO}
\title{FactorCategoryGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorCategoryGetDTO
}
\description{
FactorCategoryGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
