% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDetailsGetDTO.r
\docType{data}
\name{FactorDetailsGetDTO}
\alias{FactorDetailsGetDTO}
\title{FactorDetailsGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorDetailsGetDTO
}
\description{
FactorDetailsGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{category}}{}

\item{\code{description}}{}

\item{\code{levels}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}

\item{\code{experiment}}{}
}}

\keyword{datasets}
