% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorGetDTO.r
\docType{data}
\name{FactorGetDTO}
\alias{FactorGetDTO}
\title{FactorGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorGetDTO
}
\description{
FactorGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{category}}{}

\item{\code{description}}{}

\item{\code{experiment}}{}
}}

\keyword{datasets}
