% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorLevelGetDetailDTO.r
\docType{data}
\name{FactorLevelGetDetailDTO}
\alias{FactorLevelGetDetailDTO}
\title{FactorLevelGetDetailDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorLevelGetDetailDTO
}
\description{
FactorLevelGetDetailDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{factor}}{}

\item{\code{factor_name}}{}
}}

\keyword{datasets}
