% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorsApi.r
\name{FactorsApi}
\alias{FactorsApi}
\title{Factors operations}
\description{
opensilexClientToolsR.Factors
}
\section{Methods}{

\describe{

create_factor Create a factor


delete_factor Delete a factor


delete_factor_level Delete a factor level


get_factor_associated_experiments Get factor associated experiments


get_factor_by_uri Get a factor


get_factor_level Get a factor level


get_factor_level_detail Get a factor level


get_factor_levels Get factor levels


get_factors_by_ur_is Get a list of factors by their URIs


search_categories Search categories


search_factor_levels Search factors levels


search_factors Search factors


update_factor Update a factor

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FactorsApi$new()}}
\item \href{#method-create_factor}{\code{FactorsApi$create_factor()}}
\item \href{#method-delete_factor}{\code{FactorsApi$delete_factor()}}
\item \href{#method-delete_factor_level}{\code{FactorsApi$delete_factor_level()}}
\item \href{#method-get_factor_associated_experiments}{\code{FactorsApi$get_factor_associated_experiments()}}
\item \href{#method-get_factor_by_uri}{\code{FactorsApi$get_factor_by_uri()}}
\item \href{#method-get_factor_level}{\code{FactorsApi$get_factor_level()}}
\item \href{#method-get_factor_level_detail}{\code{FactorsApi$get_factor_level_detail()}}
\item \href{#method-get_factor_levels}{\code{FactorsApi$get_factor_levels()}}
\item \href{#method-get_factors_by_ur_is}{\code{FactorsApi$get_factors_by_ur_is()}}
\item \href{#method-search_categories}{\code{FactorsApi$search_categories()}}
\item \href{#method-search_factor_levels}{\code{FactorsApi$search_factor_levels()}}
\item \href{#method-search_factors}{\code{FactorsApi$search_factors()}}
\item \href{#method-update_factor}{\code{FactorsApi$update_factor()}}
\item \href{#method-clone}{\code{FactorsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_factor"></a>}}
\if{latex}{\out{\hypertarget{method-create_factor}{}}}
\subsection{Method \code{create_factor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$create_factor(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_factor"></a>}}
\if{latex}{\out{\hypertarget{method-delete_factor}{}}}
\subsection{Method \code{delete_factor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$delete_factor(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_factor_level"></a>}}
\if{latex}{\out{\hypertarget{method-delete_factor_level}{}}}
\subsection{Method \code{delete_factor_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$delete_factor_level(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_associated_experiments"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_associated_experiments}{}}}
\subsection{Method \code{get_factor_associated_experiments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$get_factor_associated_experiments(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_by_uri}{}}}
\subsection{Method \code{get_factor_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$get_factor_by_uri(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_level"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_level}{}}}
\subsection{Method \code{get_factor_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$get_factor_level(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_level_detail"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_level_detail}{}}}
\subsection{Method \code{get_factor_level_detail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$get_factor_level_detail(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_levels"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_levels}{}}}
\subsection{Method \code{get_factor_levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$get_factor_levels(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factors_by_ur_is"></a>}}
\if{latex}{\out{\hypertarget{method-get_factors_by_ur_is}{}}}
\subsection{Method \code{get_factors_by_ur_is()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$get_factors_by_ur_is(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_categories"></a>}}
\if{latex}{\out{\hypertarget{method-search_categories}{}}}
\subsection{Method \code{search_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$search_categories(name, order_by, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_factor_levels"></a>}}
\if{latex}{\out{\hypertarget{method-search_factor_levels}{}}}
\subsection{Method \code{search_factor_levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$search_factor_levels(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_factors"></a>}}
\if{latex}{\out{\hypertarget{method-search_factors}{}}}
\subsection{Method \code{search_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$search_factors(
  name,
  description,
  category,
  experiment,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_factor"></a>}}
\if{latex}{\out{\hypertarget{method-update_factor}{}}}
\subsection{Method \code{update_factor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$update_factor(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
