% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FontConfigDTO.r
\docType{data}
\name{FontConfigDTO}
\alias{FontConfigDTO}
\title{FontConfigDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FontConfigDTO
}
\description{
FontConfigDTO Class
}
\section{Fields}{

\describe{
\item{\code{family}}{}

\item{\code{style}}{}

\item{\code{weight}}{}

\item{\code{url}}{}

\item{\code{src}}{}
}}

\keyword{datasets}
