% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormCreationDTO.r
\docType{data}
\name{FormCreationDTO}
\alias{FormCreationDTO}
\title{FormCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FormCreationDTO
}
\description{
FormCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{timezone}}{}

\item{\code{modified_date}}{}

\item{\code{form_data}}{}

\item{\code{commit_address}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
