% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormGetDTO.r
\docType{data}
\name{FormGetDTO}
\alias{FormGetDTO}
\title{FormGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FormGetDTO
}
\description{
FormGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{timezone}}{}

\item{\code{uri}}{}

\item{\code{modified_date}}{}

\item{\code{form_data}}{}

\item{\code{commit_address}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
