% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontConfigDTO.r
\name{FrontConfigDTO}
\alias{FrontConfigDTO}
\title{FrontConfigDTO Class}
\description{
FrontConfigDTO Class

FrontConfigDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pathPrefix}}{}

\item{\code{homeComponent}}{}

\item{\code{notFoundComponent}}{}

\item{\code{headerComponent}}{}

\item{\code{loginComponent}}{}

\item{\code{menuComponent}}{}

\item{\code{footerComponent}}{}

\item{\code{menu}}{}

\item{\code{routes}}{}

\item{\code{themeModule}}{}

\item{\code{themeName}}{}

\item{\code{openIDAuthenticationURI}}{}

\item{\code{openIDConnectionTitle}}{}

\item{\code{activateResetPassword}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FrontConfigDTO$new()}}
\item \href{#method-toJSON}{\code{FrontConfigDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{FrontConfigDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{FrontConfigDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{FrontConfigDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FrontConfigDTO$fromJSONString()}}
\item \href{#method-clone}{\code{FrontConfigDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$new(
  pathPrefix,
  homeComponent,
  notFoundComponent,
  headerComponent,
  loginComponent,
  menuComponent,
  footerComponent,
  menu,
  routes,
  themeModule,
  themeName,
  openIDAuthenticationURI,
  openIDConnectionTitle,
  activateResetPassword
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$fromJSON(FrontConfigDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$fromJSONObject(FrontConfigDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$fromJSONString(FrontConfigDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrontConfigDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
