% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmApi.r
\docType{data}
\name{GermplasmApi}
\alias{GermplasmApi}
\title{Germplasm operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmApi
}
\description{
opensilexClientToolsR.Germplasm
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_germplasm Add a germplasm


delete_germplasm Delete a germplasm


export_germplasm export germplasm


export_germplasm_by_ur_is export germplasm by list of uris


get_germplasm Get a germplasm


get_germplasm_experiments Get experiments where a germplasm has been used


get_germplasms_by_uri Get a list of germplasms by their URIs


search_germplasm Search germplasm


update_germplasm Update a germplasm

}
}

\keyword{datasets}
