% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmDTO.r
\docType{data}
\name{GermplasmDTO}
\alias{GermplasmDTO}
\title{GermplasmDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmDTO
}
\description{
GermplasmDTO Class
}
\section{Fields}{

\describe{
\item{\code{accessionNumber}}{}

\item{\code{acquisitionDate}}{}

\item{\code{additionalInfo}}{}

\item{\code{biologicalStatusOfAccessionCode}}{}

\item{\code{biologicalStatusOfAccessionDescription}}{}

\item{\code{breedingMethodDbId}}{}

\item{\code{collection}}{}

\item{\code{commonCropName}}{}

\item{\code{countryOfOriginCode}}{}

\item{\code{defaultDisplayName}}{}

\item{\code{documentationURL}}{}

\item{\code{donors}}{}

\item{\code{externalReferences}}{}

\item{\code{genus}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{germplasmOrigin}}{}

\item{\code{germplasmPreprocessing}}{}

\item{\code{instituteCode}}{}

\item{\code{instituteName}}{}

\item{\code{pedigree}}{}

\item{\code{seedSource}}{}

\item{\code{seedSourceDescription}}{}

\item{\code{species}}{}

\item{\code{speciesAuthority}}{}

\item{\code{storageTypes}}{}

\item{\code{subtaxa}}{}

\item{\code{subtaxaAuthority}}{}

\item{\code{synonyms}}{}

\item{\code{taxonIds}}{}
}}

\keyword{datasets}
