% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmGetSingleDTO.r
\name{GermplasmGetSingleDTO}
\alias{GermplasmGetSingleDTO}
\title{GermplasmGetSingleDTO Class}
\description{
GermplasmGetSingleDTO Class

GermplasmGetSingleDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{synonyms}}{}

\item{\code{code}}{}

\item{\code{production_year}}{}

\item{\code{description}}{}

\item{\code{species}}{}

\item{\code{species_name}}{}

\item{\code{variety}}{}

\item{\code{variety_name}}{}

\item{\code{accession}}{}

\item{\code{accession_name}}{}

\item{\code{institute}}{}

\item{\code{website}}{}

\item{\code{metadata}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GermplasmGetSingleDTO$new()}}
\item \href{#method-toJSON}{\code{GermplasmGetSingleDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{GermplasmGetSingleDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{GermplasmGetSingleDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{GermplasmGetSingleDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GermplasmGetSingleDTO$fromJSONString()}}
\item \href{#method-clone}{\code{GermplasmGetSingleDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$new(
  uri,
  rdf_type,
  rdf_type_name,
  name,
  synonyms,
  code,
  production_year,
  description,
  species,
  species_name,
  variety,
  variety_name,
  accession,
  accession_name,
  institute,
  website,
  metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$fromJSON(GermplasmGetSingleDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$fromJSONObject(GermplasmGetSingleDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$fromJSONString(GermplasmGetSingleDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmGetSingleDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
