% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmUpdateDTO.r
\name{GermplasmUpdateDTO}
\alias{GermplasmUpdateDTO}
\title{GermplasmUpdateDTO Class}
\description{
GermplasmUpdateDTO Class

GermplasmUpdateDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{synonyms}}{}

\item{\code{code}}{}

\item{\code{production_year}}{}

\item{\code{description}}{}

\item{\code{species}}{}

\item{\code{variety}}{}

\item{\code{accession}}{}

\item{\code{institute}}{}

\item{\code{website}}{}

\item{\code{metadata}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GermplasmUpdateDTO$new()}}
\item \href{#method-toJSON}{\code{GermplasmUpdateDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{GermplasmUpdateDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{GermplasmUpdateDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{GermplasmUpdateDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GermplasmUpdateDTO$fromJSONString()}}
\item \href{#method-clone}{\code{GermplasmUpdateDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$new(
  uri,
  rdf_type,
  name,
  synonyms,
  code,
  production_year,
  description,
  species,
  variety,
  accession,
  institute,
  website,
  metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$fromJSON(GermplasmUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$fromJSONObject(GermplasmUpdateDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$fromJSONString(GermplasmUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmUpdateDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
