% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupUpdateDTO.r
\docType{data}
\name{GroupUpdateDTO}
\alias{GroupUpdateDTO}
\title{GroupUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupUpdateDTO
}
\description{
GroupUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{user_profiles}}{}
}}

\keyword{datasets}
