% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureFacilityGetDTO.r
\docType{data}
\name{InfrastructureFacilityGetDTO}
\alias{InfrastructureFacilityGetDTO}
\title{InfrastructureFacilityGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureFacilityGetDTO
}
\description{
InfrastructureFacilityGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{organisation}}{}
}}

\keyword{datasets}
