% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureFacilityNamedDTO.r
\docType{data}
\name{InfrastructureFacilityNamedDTO}
\alias{InfrastructureFacilityNamedDTO}
\title{InfrastructureFacilityNamedDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureFacilityNamedDTO
}
\description{
InfrastructureFacilityNamedDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
