% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureFacilityUpdateDTO.r
\name{InfrastructureFacilityUpdateDTO}
\alias{InfrastructureFacilityUpdateDTO}
\title{InfrastructureFacilityUpdateDTO Class}
\description{
InfrastructureFacilityUpdateDTO Class

InfrastructureFacilityUpdateDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{organisation}}{}

\item{\code{relations}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InfrastructureFacilityUpdateDTO$new()}}
\item \href{#method-toJSON}{\code{InfrastructureFacilityUpdateDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{InfrastructureFacilityUpdateDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{InfrastructureFacilityUpdateDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{InfrastructureFacilityUpdateDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InfrastructureFacilityUpdateDTO$fromJSONString()}}
\item \href{#method-clone}{\code{InfrastructureFacilityUpdateDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$new(
  uri,
  rdf_type,
  rdf_type_name,
  name,
  organisation,
  relations
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$fromJSON(InfrastructureFacilityUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$fromJSONObject(
  InfrastructureFacilityUpdateDTOObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$fromJSONString(
  InfrastructureFacilityUpdateDTOJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureFacilityUpdateDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
