% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureGetDTO.r
\docType{data}
\name{InfrastructureGetDTO}
\alias{InfrastructureGetDTO}
\title{InfrastructureGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureGetDTO
}
\description{
InfrastructureGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{parents}}{}

\item{\code{children}}{}

\item{\code{groups}}{}

\item{\code{facilities}}{}

\item{\code{sites}}{}

\item{\code{experiments}}{}
}}

\keyword{datasets}
