% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureTeamDTO.r
\docType{data}
\name{InfrastructureTeamDTO}
\alias{InfrastructureTeamDTO}
\title{InfrastructureTeamDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureTeamDTO
}
\description{
InfrastructureTeamDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{user_profiles}}{}

\item{\code{organisation}}{}
}}

\keyword{datasets}
