% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureUpdateDTO.r
\docType{data}
\name{InfrastructureUpdateDTO}
\alias{InfrastructureUpdateDTO}
\title{InfrastructureUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureUpdateDTO
}
\description{
InfrastructureUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{parent}}{}

\item{\code{children}}{}
}}

\keyword{datasets}
