% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterestEntityGetDTO.r
\docType{data}
\name{InterestEntityGetDTO}
\alias{InterestEntityGetDTO}
\title{InterestEntityGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InterestEntityGetDTO
}
\description{
InterestEntityGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
