% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Level.r
\docType{data}
\name{Level}
\alias{Level}
\title{Level Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Level
}
\description{
Level Class
}
\section{Fields}{

\describe{
\item{\code{levelInt}}{}

\item{\code{levelStr}}{}
}}

\keyword{datasets}
