% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListItemDTO.r
\docType{data}
\name{ListItemDTO}
\alias{ListItemDTO}
\title{ListItemDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ListItemDTO
}
\description{
ListItemDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
