% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodDetailsDTO.r
\docType{data}
\name{MethodDetailsDTO}
\alias{MethodDetailsDTO}
\title{MethodDetailsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MethodDetailsDTO
}
\description{
MethodDetailsDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
