% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MobileApi.r
\name{MobileApi}
\alias{MobileApi}
\title{Mobile operations}
\description{
opensilexClientToolsR.Mobile
}
\section{Methods}{

\describe{

create_form Add a form


delete_form Delete form


search_form_list Search forms


update1 Update form

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MobileApi$new()}}
\item \href{#method-create_form}{\code{MobileApi$create_form()}}
\item \href{#method-delete_form}{\code{MobileApi$delete_form()}}
\item \href{#method-search_form_list}{\code{MobileApi$search_form_list()}}
\item \href{#method-update1}{\code{MobileApi$update1()}}
\item \href{#method-clone}{\code{MobileApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MobileApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_form"></a>}}
\if{latex}{\out{\hypertarget{method-create_form}{}}}
\subsection{Method \code{create_form()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MobileApi$create_form(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_form"></a>}}
\if{latex}{\out{\hypertarget{method-delete_form}{}}}
\subsection{Method \code{delete_form()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MobileApi$delete_form(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_form_list"></a>}}
\if{latex}{\out{\hypertarget{method-search_form_list}{}}}
\subsection{Method \code{search_form_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MobileApi$search_form_list(uris, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update1"></a>}}
\if{latex}{\out{\hypertarget{method-update1}{}}}
\subsection{Method \code{update1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MobileApi$update1(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MobileApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
