% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotivationGetDTO.r
\docType{data}
\name{MotivationGetDTO}
\alias{MotivationGetDTO}
\title{MotivationGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MotivationGetDTO
}
\description{
MotivationGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
