% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveUpdateDTO.r
\name{MoveUpdateDTO}
\alias{MoveUpdateDTO}
\title{MoveUpdateDTO Class}
\description{
MoveUpdateDTO Class

MoveUpdateDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{relations}}{}

\item{\code{from}}{}

\item{\code{to}}{}

\item{\code{targets_positions}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MoveUpdateDTO$new()}}
\item \href{#method-toJSON}{\code{MoveUpdateDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{MoveUpdateDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{MoveUpdateDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{MoveUpdateDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MoveUpdateDTO$fromJSONString()}}
\item \href{#method-clone}{\code{MoveUpdateDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$new(
  uri,
  rdf_type,
  start,
  end,
  is_instant,
  description,
  targets,
  relations,
  from,
  to,
  targets_positions
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$fromJSON(MoveUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$fromJSONObject(MoveUpdateDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$fromJSONString(MoveUpdateDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MoveUpdateDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
