% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTO.r
\docType{data}
\name{NamedResourceDTO}
\alias{NamedResourceDTO}
\title{NamedResourceDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTO
}
\description{
NamedResourceDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
