% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOExperimentModel.r
\docType{data}
\name{NamedResourceDTOExperimentModel}
\alias{NamedResourceDTOExperimentModel}
\title{NamedResourceDTOExperimentModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOExperimentModel
}
\description{
NamedResourceDTOExperimentModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
