% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOFactorLevelModel.r
\docType{data}
\name{NamedResourceDTOFactorLevelModel}
\alias{NamedResourceDTOFactorLevelModel}
\title{NamedResourceDTOFactorLevelModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOFactorLevelModel
}
\description{
NamedResourceDTOFactorLevelModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
