% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOGroupModel.r
\docType{data}
\name{NamedResourceDTOGroupModel}
\alias{NamedResourceDTOGroupModel}
\title{NamedResourceDTOGroupModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOGroupModel
}
\description{
NamedResourceDTOGroupModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
