% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOInfrastructureFacilityModel.r
\docType{data}
\name{NamedResourceDTOInfrastructureFacilityModel}
\alias{NamedResourceDTOInfrastructureFacilityModel}
\title{NamedResourceDTOInfrastructureFacilityModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOInfrastructureFacilityModel
}
\description{
NamedResourceDTOInfrastructureFacilityModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
