% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOInterestEntityModel.r
\docType{data}
\name{NamedResourceDTOInterestEntityModel}
\alias{NamedResourceDTOInterestEntityModel}
\title{NamedResourceDTOInterestEntityModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOInterestEntityModel
}
\description{
NamedResourceDTOInterestEntityModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
