% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOProjectModel.r
\docType{data}
\name{NamedResourceDTOProjectModel}
\alias{NamedResourceDTOProjectModel}
\title{NamedResourceDTOProjectModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOProjectModel
}
\description{
NamedResourceDTOProjectModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
