% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedResourceDTOSiteModel.r
\docType{data}
\name{NamedResourceDTOSiteModel}
\alias{NamedResourceDTOSiteModel}
\title{NamedResourceDTOSiteModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedResourceDTOSiteModel
}
\description{
NamedResourceDTOSiteModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
