% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectUriResponse.r
\docType{data}
\name{ObjectUriResponse}
\alias{ObjectUriResponse}
\title{ObjectUriResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ObjectUriResponse
}
\description{
ObjectUriResponse Class
}
\section{Fields}{

\describe{
\item{\code{metadata}}{}

\item{\code{result}}{}
}}

\keyword{datasets}
