% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationDTO.r
\name{ObservationDTO}
\alias{ObservationDTO}
\title{ObservationDTO Class}
\description{
ObservationDTO Class

ObservationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{observationDbId}}{}

\item{\code{observationLevel}}{}

\item{\code{observationTimeStamp}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitName}}{}

\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{operator}}{}

\item{\code{season}}{}

\item{\code{studyDbId}}{}

\item{\code{uploadedBy}}{}

\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObservationDTO$new()}}
\item \href{#method-toJSON}{\code{ObservationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ObservationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ObservationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ObservationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ObservationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ObservationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$new(
  germplasmDbId,
  germplasmName,
  observationDbId,
  observationLevel,
  observationTimeStamp,
  observationUnitDbId,
  observationUnitName,
  observationVariableDbId,
  observationVariableName,
  operator,
  season,
  studyDbId,
  uploadedBy,
  value
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$fromJSON(ObservationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$fromJSONObject(ObservationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$fromJSONString(ObservationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
