% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationTreatment.r
\docType{data}
\name{ObservationTreatment}
\alias{ObservationTreatment}
\title{ObservationTreatment Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ObservationTreatment
}
\description{
ObservationTreatment Class
}
\section{Fields}{

\describe{
\item{\code{factor}}{}

\item{\code{modality}}{}
}}

\keyword{datasets}
