% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationUnitDTO.r
\name{ObservationUnitDTO}
\alias{ObservationUnitDTO}
\title{ObservationUnitDTO Class}
\description{
ObservationUnitDTO Class

ObservationUnitDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blockNumber}}{}

\item{\code{entryNumber}}{}

\item{\code{entryType}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{observationLevel}}{}

\item{\code{observationLevels}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitXref}}{}

\item{\code{observations}}{}

\item{\code{plantNumber}}{}

\item{\code{plotNumber}}{}

\item{\code{positionCoordinateX}}{}

\item{\code{positionCoordinateXType}}{}

\item{\code{positionCoordinateY}}{}

\item{\code{positionCoordinateYType}}{}

\item{\code{programName}}{}

\item{\code{replicate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyLocation}}{}

\item{\code{studyLocationDbId}}{}

\item{\code{studyName}}{}

\item{\code{treatments}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObservationUnitDTO$new()}}
\item \href{#method-toJSON}{\code{ObservationUnitDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ObservationUnitDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ObservationUnitDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ObservationUnitDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ObservationUnitDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ObservationUnitDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$new(
  blockNumber,
  entryNumber,
  entryType,
  germplasmDbId,
  germplasmName,
  observationLevel,
  observationLevels,
  observationUnitDbId,
  observationUnitXref,
  observations,
  plantNumber,
  plotNumber,
  positionCoordinateX,
  positionCoordinateXType,
  positionCoordinateY,
  positionCoordinateYType,
  programName,
  replicate,
  studyDbId,
  studyLocation,
  studyLocationDbId,
  studyName,
  treatments
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$fromJSON(ObservationUnitDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$fromJSONObject(ObservationUnitDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$fromJSONString(ObservationUnitDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnitDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
