% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationUnitXref.r
\docType{data}
\name{ObservationUnitXref}
\alias{ObservationUnitXref}
\title{ObservationUnitXref Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ObservationUnitXref
}
\description{
ObservationUnitXref Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{source}}{}
}}

\keyword{datasets}
