% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationVariableDTO.r
\name{ObservationVariableDTO}
\alias{ObservationVariableDTO}
\title{ObservationVariableDTO Class}
\description{
ObservationVariableDTO Class

ObservationVariableDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{ontologyReference}}{}

\item{\code{synonyms}}{}

\item{\code{contextOfUse}}{}

\item{\code{growthStage}}{}

\item{\code{status}}{}

\item{\code{xref}}{}

\item{\code{institution}}{}

\item{\code{scientist}}{}

\item{\code{submissionTimesTamp}}{}

\item{\code{language}}{}

\item{\code{crop}}{}

\item{\code{trait}}{}

\item{\code{method}}{}

\item{\code{scale}}{}

\item{\code{defaultValue}}{}

\item{\code{documentationURL}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObservationVariableDTO$new()}}
\item \href{#method-toJSON}{\code{ObservationVariableDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ObservationVariableDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ObservationVariableDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ObservationVariableDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ObservationVariableDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ObservationVariableDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$new(
  observationVariableDbId,
  observationVariableName,
  ontologyReference,
  synonyms,
  contextOfUse,
  growthStage,
  status,
  xref,
  institution,
  scientist,
  submissionTimesTamp,
  language,
  crop,
  trait,
  method,
  scale,
  defaultValue,
  documentationURL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$fromJSON(ObservationVariableDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$fromJSONObject(ObservationVariableDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$fromJSONString(ObservationVariableDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariableDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
