% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyApi.r
\name{OntologyApi}
\alias{OntologyApi}
\title{Ontology operations}
\description{
opensilexClientToolsR.Ontology
}
\section{Methods}{

\describe{

add_class_property_restriction Add a rdf type property restriction


check_ur_is_types Check the given rdf-types on the given uris


create_property Create a RDF property


delete_class_property_restriction Delete a rdf type property restriction


delete_property Delete a property


get_classes Return classes models definitions with properties for a list of rdf types


get_data_properties Search data properties tree


get_object_properties Search object properties tree


get_properties Search properties tree


get_property Return property model definition detail


get_rdf_type Return class model definition with properties


get_sub_classes_of Search sub-classes tree of an RDF class


get_uri_label Return associated rdfs:label of an uri if exists


get_uri_labels_list Return associated rdfs:label of uris if they exist


search_sub_classes_of Search sub-classes tree of an RDF class


update_class_property_restriction Update a rdf type property restriction


update_property Update a RDF property

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OntologyApi$new()}}
\item \href{#method-add_class_property_restriction}{\code{OntologyApi$add_class_property_restriction()}}
\item \href{#method-check_ur_is_types}{\code{OntologyApi$check_ur_is_types()}}
\item \href{#method-create_property}{\code{OntologyApi$create_property()}}
\item \href{#method-delete_class_property_restriction}{\code{OntologyApi$delete_class_property_restriction()}}
\item \href{#method-delete_property}{\code{OntologyApi$delete_property()}}
\item \href{#method-get_classes}{\code{OntologyApi$get_classes()}}
\item \href{#method-get_data_properties}{\code{OntologyApi$get_data_properties()}}
\item \href{#method-get_object_properties}{\code{OntologyApi$get_object_properties()}}
\item \href{#method-get_properties}{\code{OntologyApi$get_properties()}}
\item \href{#method-get_property}{\code{OntologyApi$get_property()}}
\item \href{#method-get_rdf_type}{\code{OntologyApi$get_rdf_type()}}
\item \href{#method-get_sub_classes_of}{\code{OntologyApi$get_sub_classes_of()}}
\item \href{#method-get_uri_label}{\code{OntologyApi$get_uri_label()}}
\item \href{#method-get_uri_labels_list}{\code{OntologyApi$get_uri_labels_list()}}
\item \href{#method-search_sub_classes_of}{\code{OntologyApi$search_sub_classes_of()}}
\item \href{#method-update_class_property_restriction}{\code{OntologyApi$update_class_property_restriction()}}
\item \href{#method-update_property}{\code{OntologyApi$update_property()}}
\item \href{#method-clone}{\code{OntologyApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_class_property_restriction"></a>}}
\if{latex}{\out{\hypertarget{method-add_class_property_restriction}{}}}
\subsection{Method \code{add_class_property_restriction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$add_class_property_restriction(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_ur_is_types"></a>}}
\if{latex}{\out{\hypertarget{method-check_ur_is_types}{}}}
\subsection{Method \code{check_ur_is_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$check_ur_is_types(rdf_types, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_property"></a>}}
\if{latex}{\out{\hypertarget{method-create_property}{}}}
\subsection{Method \code{create_property()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$create_property(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_class_property_restriction"></a>}}
\if{latex}{\out{\hypertarget{method-delete_class_property_restriction}{}}}
\subsection{Method \code{delete_class_property_restriction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$delete_class_property_restriction(rdf_type, property_uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_property"></a>}}
\if{latex}{\out{\hypertarget{method-delete_property}{}}}
\subsection{Method \code{delete_property()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$delete_property(property_uri, property_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_classes"></a>}}
\if{latex}{\out{\hypertarget{method-get_classes}{}}}
\subsection{Method \code{get_classes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_classes(rdf_type, parent_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_properties"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_properties}{}}}
\subsection{Method \code{get_data_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_data_properties(domain, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_object_properties"></a>}}
\if{latex}{\out{\hypertarget{method-get_object_properties}{}}}
\subsection{Method \code{get_object_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_object_properties(domain, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_properties"></a>}}
\if{latex}{\out{\hypertarget{method-get_properties}{}}}
\subsection{Method \code{get_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_properties(domain, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_property"></a>}}
\if{latex}{\out{\hypertarget{method-get_property}{}}}
\subsection{Method \code{get_property()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_property(uri, rdf_type, domain_rdf_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rdf_type"></a>}}
\if{latex}{\out{\hypertarget{method-get_rdf_type}{}}}
\subsection{Method \code{get_rdf_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_rdf_type(rdf_type, parent_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sub_classes_of"></a>}}
\if{latex}{\out{\hypertarget{method-get_sub_classes_of}{}}}
\subsection{Method \code{get_sub_classes_of()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_sub_classes_of(parent_type, ignore_root_classes, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uri_label"></a>}}
\if{latex}{\out{\hypertarget{method-get_uri_label}{}}}
\subsection{Method \code{get_uri_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_uri_label(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uri_labels_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_uri_labels_list}{}}}
\subsection{Method \code{get_uri_labels_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$get_uri_labels_list(uri, context, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_sub_classes_of"></a>}}
\if{latex}{\out{\hypertarget{method-search_sub_classes_of}{}}}
\subsection{Method \code{search_sub_classes_of()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$search_sub_classes_of(parent_type, name, ignore_root_classes, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_class_property_restriction"></a>}}
\if{latex}{\out{\hypertarget{method-update_class_property_restriction}{}}}
\subsection{Method \code{update_class_property_restriction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$update_class_property_restriction(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_property"></a>}}
\if{latex}{\out{\hypertarget{method-update_property}{}}}
\subsection{Method \code{update_property()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$update_property(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
