% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderBy.r
\docType{data}
\name{OrderBy}
\alias{OrderBy}
\title{OrderBy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OrderBy
}
\description{
OrderBy Class
}
\section{Fields}{

\describe{
\item{\code{fieldName}}{}

\item{\code{order}}{}
}}

\keyword{datasets}
