% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganisationsApi.r
\name{OrganisationsApi}
\alias{OrganisationsApi}
\title{Organisations operations}
\description{
opensilexClientToolsR.Organisations
}
\section{Methods}{

\describe{

create_infrastructure Create an organisation


create_infrastructure_facility Create a facility


create_infrastructure_team Create a team


delete_infrastructure Delete an organisation


delete_infrastructure_facility Delete a facility


delete_infrastructure_team Delete a team


get_all_facilities Get all facilities


get_facilities_by_uri Get facilities by their URIs


get_infrastructure Get an organisation 


get_infrastructure_facility Get a facility


get_infrastructure_team Get a team


search_infrastructure_facilities Search facilities


search_infrastructures_tree Search organisations


update_infrastructure Update an organisation


update_infrastructure_facility Update a facility


update_infrastructure_team Update a team

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrganisationsApi$new()}}
\item \href{#method-create_infrastructure}{\code{OrganisationsApi$create_infrastructure()}}
\item \href{#method-create_infrastructure_facility}{\code{OrganisationsApi$create_infrastructure_facility()}}
\item \href{#method-create_infrastructure_team}{\code{OrganisationsApi$create_infrastructure_team()}}
\item \href{#method-delete_infrastructure}{\code{OrganisationsApi$delete_infrastructure()}}
\item \href{#method-delete_infrastructure_facility}{\code{OrganisationsApi$delete_infrastructure_facility()}}
\item \href{#method-delete_infrastructure_team}{\code{OrganisationsApi$delete_infrastructure_team()}}
\item \href{#method-get_all_facilities}{\code{OrganisationsApi$get_all_facilities()}}
\item \href{#method-get_facilities_by_uri}{\code{OrganisationsApi$get_facilities_by_uri()}}
\item \href{#method-get_infrastructure}{\code{OrganisationsApi$get_infrastructure()}}
\item \href{#method-get_infrastructure_facility}{\code{OrganisationsApi$get_infrastructure_facility()}}
\item \href{#method-get_infrastructure_team}{\code{OrganisationsApi$get_infrastructure_team()}}
\item \href{#method-search_infrastructure_facilities}{\code{OrganisationsApi$search_infrastructure_facilities()}}
\item \href{#method-search_infrastructures_tree}{\code{OrganisationsApi$search_infrastructures_tree()}}
\item \href{#method-update_infrastructure}{\code{OrganisationsApi$update_infrastructure()}}
\item \href{#method-update_infrastructure_facility}{\code{OrganisationsApi$update_infrastructure_facility()}}
\item \href{#method-update_infrastructure_team}{\code{OrganisationsApi$update_infrastructure_team()}}
\item \href{#method-clone}{\code{OrganisationsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_infrastructure"></a>}}
\if{latex}{\out{\hypertarget{method-create_infrastructure}{}}}
\subsection{Method \code{create_infrastructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$create_infrastructure(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_infrastructure_facility"></a>}}
\if{latex}{\out{\hypertarget{method-create_infrastructure_facility}{}}}
\subsection{Method \code{create_infrastructure_facility()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$create_infrastructure_facility(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_infrastructure_team"></a>}}
\if{latex}{\out{\hypertarget{method-create_infrastructure_team}{}}}
\subsection{Method \code{create_infrastructure_team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$create_infrastructure_team(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_infrastructure"></a>}}
\if{latex}{\out{\hypertarget{method-delete_infrastructure}{}}}
\subsection{Method \code{delete_infrastructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$delete_infrastructure(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_infrastructure_facility"></a>}}
\if{latex}{\out{\hypertarget{method-delete_infrastructure_facility}{}}}
\subsection{Method \code{delete_infrastructure_facility()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$delete_infrastructure_facility(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_infrastructure_team"></a>}}
\if{latex}{\out{\hypertarget{method-delete_infrastructure_team}{}}}
\subsection{Method \code{delete_infrastructure_team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$delete_infrastructure_team(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_facilities"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_facilities}{}}}
\subsection{Method \code{get_all_facilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$get_all_facilities(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_facilities_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_facilities_by_uri}{}}}
\subsection{Method \code{get_facilities_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$get_facilities_by_uri(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_infrastructure"></a>}}
\if{latex}{\out{\hypertarget{method-get_infrastructure}{}}}
\subsection{Method \code{get_infrastructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$get_infrastructure(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_infrastructure_facility"></a>}}
\if{latex}{\out{\hypertarget{method-get_infrastructure_facility}{}}}
\subsection{Method \code{get_infrastructure_facility()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$get_infrastructure_facility(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_infrastructure_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_infrastructure_team}{}}}
\subsection{Method \code{get_infrastructure_team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$get_infrastructure_team(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_infrastructure_facilities"></a>}}
\if{latex}{\out{\hypertarget{method-search_infrastructure_facilities}{}}}
\subsection{Method \code{search_infrastructure_facilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$search_infrastructure_facilities(
  pattern,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_infrastructures_tree"></a>}}
\if{latex}{\out{\hypertarget{method-search_infrastructures_tree}{}}}
\subsection{Method \code{search_infrastructures_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$search_infrastructures_tree(pattern, organisation_uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_infrastructure"></a>}}
\if{latex}{\out{\hypertarget{method-update_infrastructure}{}}}
\subsection{Method \code{update_infrastructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$update_infrastructure(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_infrastructure_facility"></a>}}
\if{latex}{\out{\hypertarget{method-update_infrastructure_facility}{}}}
\subsection{Method \code{update_infrastructure_facility()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$update_infrastructure_facility(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_infrastructure_team"></a>}}
\if{latex}{\out{\hypertarget{method-update_infrastructure_team}{}}}
\subsection{Method \code{update_infrastructure_team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$update_infrastructure_team(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
