% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizationsApi.r
\docType{data}
\name{OrganizationsApi}
\alias{OrganizationsApi}
\title{Organizations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OrganizationsApi
}
\description{
opensilexClientToolsR.Organizations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_infrastructure Create an organisation


create_infrastructure_facility Create a facility


create_site Create a site


delete_infrastructure Delete an organisation


delete_infrastructure_facility Delete a facility


delete_site Delete a site


get_all_facilities Get all facilities


get_facilities_by_uri Get facilities by their URIs


get_infrastructure Get an organisation 


get_infrastructure_facility Get a facility


get_site Get a site


get_sites_by_uri Get a list of sites


search_infrastructure_facilities Search facilities


search_infrastructures Search organisations


search_sites Search all sites


update_infrastructure Update an organisation


update_infrastructure_facility Update a facility


update_site Update a site

}
}

\keyword{datasets}
