% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWLClassPropertyRestrictionDTO.r
\docType{data}
\name{OWLClassPropertyRestrictionDTO}
\alias{OWLClassPropertyRestrictionDTO}
\title{OWLClassPropertyRestrictionDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OWLClassPropertyRestrictionDTO
}
\description{
OWLClassPropertyRestrictionDTO Class
}
\section{Fields}{

\describe{
\item{\code{property}}{}

\item{\code{required}}{}

\item{\code{list}}{}

\item{\code{rdf_type}}{}

\item{\code{domain}}{}
}}

\keyword{datasets}
