% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaginationDTO.r
\docType{data}
\name{PaginationDTO}
\alias{PaginationDTO}
\title{PaginationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PaginationDTO
}
\description{
PaginationDTO Class
}
\section{Fields}{

\describe{
\item{\code{pageSize}}{}

\item{\code{currentPage}}{}

\item{\code{totalCount}}{}

\item{\code{totalPages}}{}
}}

\keyword{datasets}
