% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionCreationDTO.r
\docType{data}
\name{PositionCreationDTO}
\alias{PositionCreationDTO}
\title{PositionCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PositionCreationDTO
}
\description{
PositionCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{point}}{}

\item{\code{x}}{}

\item{\code{y}}{}

\item{\code{z}}{}

\item{\code{text}}{}
}}

\keyword{datasets}
