% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionGetDetailDTO.r
\docType{data}
\name{PositionGetDetailDTO}
\alias{PositionGetDetailDTO}
\title{PositionGetDetailDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PositionGetDetailDTO
}
\description{
PositionGetDetailDTO Class
}
\section{Fields}{

\describe{
\item{\code{point}}{}

\item{\code{x}}{}

\item{\code{y}}{}

\item{\code{z}}{}

\item{\code{text}}{}
}}

\keyword{datasets}
