% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionGetDTO.r
\docType{data}
\name{PositionGetDTO}
\alias{PositionGetDTO}
\title{PositionGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PositionGetDTO
}
\description{
PositionGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{event}}{}

\item{\code{move_time}}{}

\item{\code{from}}{}

\item{\code{to}}{}

\item{\code{position}}{}
}}

\keyword{datasets}
