% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionsApi.r
\name{PositionsApi}
\alias{PositionsApi}
\title{Positions operations}
\description{
opensilexClientToolsR.Positions
}
\section{Methods}{

\describe{

get_position Get the position of an object


search_position_history Search history of position of an object

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PositionsApi$new()}}
\item \href{#method-get_position}{\code{PositionsApi$get_position()}}
\item \href{#method-search_position_history}{\code{PositionsApi$search_position_history()}}
\item \href{#method-clone}{\code{PositionsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-get_position}{}}}
\subsection{Method \code{get_position()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionsApi$get_position(uri, time, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_position_history"></a>}}
\if{latex}{\out{\hypertarget{method-search_position_history}{}}}
\subsection{Method \code{search_position_history()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionsApi$search_position_history(
  target,
  start_date_time,
  end_date_time,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
