% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileGetDTO.r
\docType{data}
\name{ProfileGetDTO}
\alias{ProfileGetDTO}
\title{ProfileGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProfileGetDTO
}
\description{
ProfileGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{credentials}}{}
}}

\keyword{datasets}
