% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileUpdateDTO.r
\docType{data}
\name{ProfileUpdateDTO}
\alias{ProfileUpdateDTO}
\title{ProfileUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProfileUpdateDTO
}
\description{
ProfileUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{credentials}}{}
}}

\keyword{datasets}
