% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectCreationDTO.r
\name{ProjectCreationDTO}
\alias{ProjectCreationDTO}
\title{ProjectCreationDTO Class}
\description{
ProjectCreationDTO Class

ProjectCreationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{description}}{}

\item{\code{objective}}{}

\item{\code{financial_funding}}{}

\item{\code{website}}{}

\item{\code{related_projects}}{}

\item{\code{coordinators}}{}

\item{\code{scientific_contacts}}{}

\item{\code{administrative_contacts}}{}

\item{\code{experiments}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectCreationDTO$new()}}
\item \href{#method-toJSON}{\code{ProjectCreationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectCreationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ProjectCreationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ProjectCreationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectCreationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectCreationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$new(
  uri,
  name,
  shortname,
  start_date,
  end_date,
  description,
  objective,
  financial_funding,
  website,
  related_projects,
  coordinators,
  scientific_contacts,
  administrative_contacts,
  experiments
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$fromJSON(ProjectCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$fromJSONObject(ProjectCreationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$fromJSONString(ProjectCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
