% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectGetDTO.r
\name{ProjectGetDTO}
\alias{ProjectGetDTO}
\title{ProjectGetDTO Class}
\description{
ProjectGetDTO Class

ProjectGetDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{description}}{}

\item{\code{objective}}{}

\item{\code{financial_funding}}{}

\item{\code{website}}{}

\item{\code{related_projects}}{}

\item{\code{coordinators}}{}

\item{\code{scientific_contacts}}{}

\item{\code{administrative_contacts}}{}

\item{\code{experiments}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectGetDTO$new()}}
\item \href{#method-toJSON}{\code{ProjectGetDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectGetDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ProjectGetDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ProjectGetDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectGetDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectGetDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$new(
  uri,
  name,
  shortname,
  start_date,
  end_date,
  description,
  objective,
  financial_funding,
  website,
  related_projects,
  coordinators,
  scientific_contacts,
  administrative_contacts,
  experiments
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$fromJSON(ProjectGetDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$fromJSONObject(ProjectGetDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$fromJSONString(ProjectGetDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectGetDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
