% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectsApi.r
\name{ProjectsApi}
\alias{ProjectsApi}
\title{Projects operations}
\description{
opensilexClientToolsR.Projects
}
\section{Methods}{

\describe{

create_project Add a project


delete_project Delete a project


get_project Get a project


get_projects_by_uri Get projects by their URIs


search_projects Search projects


update_project Update a project

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectsApi$new()}}
\item \href{#method-create_project}{\code{ProjectsApi$create_project()}}
\item \href{#method-delete_project}{\code{ProjectsApi$delete_project()}}
\item \href{#method-get_project}{\code{ProjectsApi$get_project()}}
\item \href{#method-get_projects_by_uri}{\code{ProjectsApi$get_projects_by_uri()}}
\item \href{#method-search_projects}{\code{ProjectsApi$search_projects()}}
\item \href{#method-update_project}{\code{ProjectsApi$update_project()}}
\item \href{#method-clone}{\code{ProjectsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_project"></a>}}
\if{latex}{\out{\hypertarget{method-create_project}{}}}
\subsection{Method \code{create_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$create_project(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_project"></a>}}
\if{latex}{\out{\hypertarget{method-delete_project}{}}}
\subsection{Method \code{delete_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$delete_project(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_project"></a>}}
\if{latex}{\out{\hypertarget{method-get_project}{}}}
\subsection{Method \code{get_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$get_project(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_projects_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_projects_by_uri}{}}}
\subsection{Method \code{get_projects_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$get_projects_by_uri(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_projects"></a>}}
\if{latex}{\out{\hypertarget{method-search_projects}{}}}
\subsection{Method \code{search_projects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$search_projects(
  name,
  year,
  keyword,
  financial_funding,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_project"></a>}}
\if{latex}{\out{\hypertarget{method-update_project}{}}}
\subsection{Method \code{update_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$update_project(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
