% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvenanceUpdateDTO.r
\docType{data}
\name{ProvenanceUpdateDTO}
\alias{ProvenanceUpdateDTO}
\title{ProvenanceUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProvenanceUpdateDTO
}
\description{
ProvenanceUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{prov_activity}}{}

\item{\code{prov_agent}}{}
}}

\keyword{datasets}
