% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvEntityModel.r
\docType{data}
\name{ProvEntityModel}
\alias{ProvEntityModel}
\title{ProvEntityModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProvEntityModel
}
\description{
ProvEntityModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}
}}

\keyword{datasets}
