% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDFPropertyDTO.r
\docType{data}
\name{RDFPropertyDTO}
\alias{RDFPropertyDTO}
\title{RDFPropertyDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RDFPropertyDTO
}
\description{
RDFPropertyDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{domain}}{}

\item{\code{range}}{}

\item{\code{parent}}{}

\item{\code{rdf_type}}{}

\item{\code{name_translations}}{}

\item{\code{comment_translations}}{}

\item{\code{domain_rdf_type}}{}

\item{\code{range_label}}{}
}}

\keyword{datasets}
