% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceDagDTO.r
\docType{data}
\name{ResourceDagDTO}
\alias{ResourceDagDTO}
\title{ResourceDagDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResourceDagDTO
}
\description{
ResourceDagDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{parents}}{}

\item{\code{children}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
