% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceTreeDTO.r
\docType{data}
\name{ResourceTreeDTO}
\alias{ResourceTreeDTO}
\title{ResourceTreeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResourceTreeDTO
}
\description{
ResourceTreeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{parent}}{}

\item{\code{selected}}{}

\item{\code{disabled}}{}

\item{\code{children}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
