% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDTO.r
\docType{data}
\name{RouteDTO}
\alias{RouteDTO}
\title{RouteDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RouteDTO
}
\description{
RouteDTO Class
}
\section{Fields}{

\describe{
\item{\code{path}}{}

\item{\code{component}}{}

\item{\code{credentials}}{}
}}

\keyword{datasets}
