% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale.r
\docType{data}
\name{Scale}
\alias{Scale}
\title{Scale Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Scale
}
\description{
Scale Class
}
\section{Fields}{

\describe{
\item{\code{dataType}}{}

\item{\code{decimalPlaces}}{}

\item{\code{name}}{}

\item{\code{ontologyReference}}{}

\item{\code{scaleDbId}}{}

\item{\code{scaleName}}{}

\item{\code{validValues}}{}

\item{\code{xref}}{}
}}

\keyword{datasets}
