% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectCsvExportDTO.r
\docType{data}
\name{ScientificObjectCsvExportDTO}
\alias{ScientificObjectCsvExportDTO}
\title{ScientificObjectCsvExportDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectCsvExportDTO
}
\description{
ScientificObjectCsvExportDTO Class
}
\section{Fields}{

\describe{
\item{\code{uris}}{}

\item{\code{experiment}}{}
}}

\keyword{datasets}
