% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectDetailByExperimentsDTO.r
\name{ScientificObjectDetailByExperimentsDTO}
\alias{ScientificObjectDetailByExperimentsDTO}
\title{ScientificObjectDetailByExperimentsDTO Class}
\description{
ScientificObjectDetailByExperimentsDTO Class

ScientificObjectDetailByExperimentsDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{parent}}{}

\item{\code{parent_name}}{}

\item{\code{experiment}}{}

\item{\code{experiment_name}}{}

\item{\code{factor_level}}{}

\item{\code{relations}}{}

\item{\code{geometry}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectDetailByExperimentsDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectDetailByExperimentsDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectDetailByExperimentsDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectDetailByExperimentsDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectDetailByExperimentsDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectDetailByExperimentsDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectDetailByExperimentsDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$new(
  uri,
  rdf_type,
  rdf_type_name,
  name,
  parent,
  parent_name,
  experiment,
  experiment_name,
  factor_level,
  relations,
  geometry
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$fromJSON(
  ScientificObjectDetailByExperimentsDTOJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$fromJSONObject(
  ScientificObjectDetailByExperimentsDTOObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$fromJSONString(
  ScientificObjectDetailByExperimentsDTOJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDetailByExperimentsDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
