% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectNodeDTO.r
\docType{data}
\name{ScientificObjectNodeDTO}
\alias{ScientificObjectNodeDTO}
\title{ScientificObjectNodeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectNodeDTO
}
\description{
ScientificObjectNodeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{geometry}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{creation_date}}{}

\item{\code{destruction_date}}{}
}}

\keyword{datasets}
