% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectNodeWithChildrenDTO.r
\name{ScientificObjectNodeWithChildrenDTO}
\alias{ScientificObjectNodeWithChildrenDTO}
\title{ScientificObjectNodeWithChildrenDTO Class}
\description{
ScientificObjectNodeWithChildrenDTO Class

ScientificObjectNodeWithChildrenDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{geometry}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{creation_date}}{}

\item{\code{destruction_date}}{}

\item{\code{child_count}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectNodeWithChildrenDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectNodeWithChildrenDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectNodeWithChildrenDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectNodeWithChildrenDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectNodeWithChildrenDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectNodeWithChildrenDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectNodeWithChildrenDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$new(
  uri,
  name,
  geometry,
  rdf_type,
  rdf_type_name,
  creation_date,
  destruction_date,
  child_count
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$fromJSON(
  ScientificObjectNodeWithChildrenDTOJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$fromJSONObject(
  ScientificObjectNodeWithChildrenDTOObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$fromJSONString(
  ScientificObjectNodeWithChildrenDTOJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeWithChildrenDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
