% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectSearchDTO.r
\name{ScientificObjectSearchDTO}
\alias{ScientificObjectSearchDTO}
\title{ScientificObjectSearchDTO Class}
\description{
ScientificObjectSearchDTO Class

ScientificObjectSearchDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uris}}{}

\item{\code{excluded_uris}}{}

\item{\code{experiment}}{}

\item{\code{rdf_types}}{}

\item{\code{name}}{}

\item{\code{parent}}{}

\item{\code{germplasm}}{}

\item{\code{factor_levels}}{}

\item{\code{facility}}{}

\item{\code{existence_date}}{}

\item{\code{creation_date}}{}

\item{\code{order_by}}{}

\item{\code{page}}{}

\item{\code{page_sze}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectSearchDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectSearchDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectSearchDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectSearchDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectSearchDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectSearchDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectSearchDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$new(
  uris,
  excluded_uris,
  experiment,
  rdf_types,
  name,
  parent,
  germplasm,
  factor_levels,
  facility,
  existence_date,
  creation_date,
  order_by,
  page,
  page_sze
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$fromJSON(ScientificObjectSearchDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$fromJSONObject(ScientificObjectSearchDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$fromJSONString(ScientificObjectSearchDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectSearchDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
