% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecurityApi.r
\name{SecurityApi}
\alias{SecurityApi}
\title{Security operations}
\description{
opensilexClientToolsR.Security
}
\section{Methods}{

\describe{

create_group Add a group


create_profile Add a profile


create_user Add an user


delete_group Delete a group


delete_profile Delete a profile


delete_user Delete an user


get_all_profiles Get all profiles


get_group Get a group


get_groups_by_uri Get groups by their URIs


get_profile Get a profile


get_user Get an user


get_user_groups Get groups of an user


get_users_by_uri Get users by their URIs


search_groups Search groups


search_profiles Search profiles


search_users Search users


update_group Update a group


update_profile Update a profile


update_user Update an user

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SecurityApi$new()}}
\item \href{#method-create_group}{\code{SecurityApi$create_group()}}
\item \href{#method-create_profile}{\code{SecurityApi$create_profile()}}
\item \href{#method-create_user}{\code{SecurityApi$create_user()}}
\item \href{#method-delete_group}{\code{SecurityApi$delete_group()}}
\item \href{#method-delete_profile}{\code{SecurityApi$delete_profile()}}
\item \href{#method-delete_user}{\code{SecurityApi$delete_user()}}
\item \href{#method-get_all_profiles}{\code{SecurityApi$get_all_profiles()}}
\item \href{#method-get_group}{\code{SecurityApi$get_group()}}
\item \href{#method-get_groups_by_uri}{\code{SecurityApi$get_groups_by_uri()}}
\item \href{#method-get_profile}{\code{SecurityApi$get_profile()}}
\item \href{#method-get_user}{\code{SecurityApi$get_user()}}
\item \href{#method-get_user_groups}{\code{SecurityApi$get_user_groups()}}
\item \href{#method-get_users_by_uri}{\code{SecurityApi$get_users_by_uri()}}
\item \href{#method-search_groups}{\code{SecurityApi$search_groups()}}
\item \href{#method-search_profiles}{\code{SecurityApi$search_profiles()}}
\item \href{#method-search_users}{\code{SecurityApi$search_users()}}
\item \href{#method-update_group}{\code{SecurityApi$update_group()}}
\item \href{#method-update_profile}{\code{SecurityApi$update_profile()}}
\item \href{#method-update_user}{\code{SecurityApi$update_user()}}
\item \href{#method-clone}{\code{SecurityApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_group"></a>}}
\if{latex}{\out{\hypertarget{method-create_group}{}}}
\subsection{Method \code{create_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$create_group(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_profile"></a>}}
\if{latex}{\out{\hypertarget{method-create_profile}{}}}
\subsection{Method \code{create_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$create_profile(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user"></a>}}
\if{latex}{\out{\hypertarget{method-create_user}{}}}
\subsection{Method \code{create_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$create_user(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_group"></a>}}
\if{latex}{\out{\hypertarget{method-delete_group}{}}}
\subsection{Method \code{delete_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$delete_group(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_profile"></a>}}
\if{latex}{\out{\hypertarget{method-delete_profile}{}}}
\subsection{Method \code{delete_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$delete_profile(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user}{}}}
\subsection{Method \code{delete_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$delete_user(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_profiles"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_profiles}{}}}
\subsection{Method \code{get_all_profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_all_profiles(order_by, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_group"></a>}}
\if{latex}{\out{\hypertarget{method-get_group}{}}}
\subsection{Method \code{get_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_group(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_groups_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_groups_by_uri}{}}}
\subsection{Method \code{get_groups_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_groups_by_uri(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_profile"></a>}}
\if{latex}{\out{\hypertarget{method-get_profile}{}}}
\subsection{Method \code{get_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_profile(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user"></a>}}
\if{latex}{\out{\hypertarget{method-get_user}{}}}
\subsection{Method \code{get_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_user(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user_groups"></a>}}
\if{latex}{\out{\hypertarget{method-get_user_groups}{}}}
\subsection{Method \code{get_user_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_user_groups(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_users_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_users_by_uri}{}}}
\subsection{Method \code{get_users_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$get_users_by_uri(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_groups"></a>}}
\if{latex}{\out{\hypertarget{method-search_groups}{}}}
\subsection{Method \code{search_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$search_groups(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_profiles"></a>}}
\if{latex}{\out{\hypertarget{method-search_profiles}{}}}
\subsection{Method \code{search_profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$search_profiles(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_users"></a>}}
\if{latex}{\out{\hypertarget{method-search_users}{}}}
\subsection{Method \code{search_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$search_users(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_group"></a>}}
\if{latex}{\out{\hypertarget{method-update_group}{}}}
\subsection{Method \code{update_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$update_group(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_profile"></a>}}
\if{latex}{\out{\hypertarget{method-update_profile}{}}}
\subsection{Method \code{update_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$update_profile(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_user"></a>}}
\if{latex}{\out{\hypertarget{method-update_user}{}}}
\subsection{Method \code{update_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$update_user(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
