% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteCreationDTO.r
\docType{data}
\name{SiteCreationDTO}
\alias{SiteCreationDTO}
\title{SiteCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SiteCreationDTO
}
\description{
SiteCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{address}}{}

\item{\code{organizations}}{}

\item{\code{facilities}}{}

\item{\code{groups}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
