% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteGetDTO.r
\docType{data}
\name{SiteGetDTO}
\alias{SiteGetDTO}
\title{SiteGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SiteGetDTO
}
\description{
SiteGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{address}}{}

\item{\code{organizations}}{}

\item{\code{facilities}}{}

\item{\code{groups}}{}

\item{\code{geometry}}{}
}}

\keyword{datasets}
