% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteUpdateDTO.r
\docType{data}
\name{SiteUpdateDTO}
\alias{SiteUpdateDTO}
\title{SiteUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SiteUpdateDTO
}
\description{
SiteUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{address}}{}

\item{\code{organizations}}{}

\item{\code{facilities}}{}

\item{\code{groups}}{}

\item{\code{rdf_type_name}}{}
}}

\keyword{datasets}
