% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesDTO.r
\docType{data}
\name{SpeciesDTO}
\alias{SpeciesDTO}
\title{SpeciesDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpeciesDTO
}
\description{
SpeciesDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
