% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatusDTO.r
\docType{data}
\name{StatusDTO}
\alias{StatusDTO}
\title{StatusDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StatusDTO
}
\description{
StatusDTO Class
}
\section{Fields}{

\describe{
\item{\code{message}}{}

\item{\code{messageType}}{}
}}

\keyword{datasets}
