% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyDetailsDTO.r
\name{StudyDetailsDTO}
\alias{StudyDetailsDTO}
\title{StudyDetailsDTO Class}
\description{
StudyDetailsDTO Class

StudyDetailsDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{active}}{}

\item{\code{additionalInfo}}{}

\item{\code{commonCropName}}{}

\item{\code{documentationURL}}{}

\item{\code{endDate}}{}

\item{\code{locationDbId}}{}

\item{\code{locationName}}{}

\item{\code{name}}{}

\item{\code{programDbId}}{}

\item{\code{programName}}{}

\item{\code{seasons}}{}

\item{\code{startDate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{studyType}}{}

\item{\code{studyTypeDbId}}{}

\item{\code{studyTypeName}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}

\item{\code{contacts}}{}

\item{\code{datalinks}}{}

\item{\code{location}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudyDetailsDTO$new()}}
\item \href{#method-toJSON}{\code{StudyDetailsDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{StudyDetailsDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{StudyDetailsDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{StudyDetailsDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StudyDetailsDTO$fromJSONString()}}
\item \href{#method-clone}{\code{StudyDetailsDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$new(
  active,
  additionalInfo,
  commonCropName,
  documentationURL,
  endDate,
  locationDbId,
  locationName,
  name,
  programDbId,
  programName,
  seasons,
  startDate,
  studyDbId,
  studyName,
  studyType,
  studyTypeDbId,
  studyTypeName,
  trialDbId,
  trialName,
  contacts,
  datalinks,
  location
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$fromJSON(StudyDetailsDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$fromJSONObject(StudyDetailsDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$fromJSONString(StudyDetailsDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetailsDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
