% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyDTO.r
\name{StudyDTO}
\alias{StudyDTO}
\title{StudyDTO Class}
\description{
StudyDTO Class

StudyDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{active}}{}

\item{\code{additionalInfo}}{}

\item{\code{commonCropName}}{}

\item{\code{documentationURL}}{}

\item{\code{endDate}}{}

\item{\code{locationDbId}}{}

\item{\code{locationName}}{}

\item{\code{name}}{}

\item{\code{programDbId}}{}

\item{\code{programName}}{}

\item{\code{seasons}}{}

\item{\code{startDate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{studyType}}{}

\item{\code{studyTypeDbId}}{}

\item{\code{studyTypeName}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudyDTO$new()}}
\item \href{#method-toJSON}{\code{StudyDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{StudyDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{StudyDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{StudyDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StudyDTO$fromJSONString()}}
\item \href{#method-clone}{\code{StudyDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$new(
  active,
  additionalInfo,
  commonCropName,
  documentationURL,
  endDate,
  locationDbId,
  locationName,
  name,
  programDbId,
  programName,
  seasons,
  startDate,
  studyDbId,
  studyName,
  studyType,
  studyTypeDbId,
  studyTypeName,
  trialDbId,
  trialName
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$fromJSON(StudyDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$fromJSONObject(StudyDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$fromJSONString(StudyDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
