% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetPositionCreationDTO.r
\docType{data}
\name{TargetPositionCreationDTO}
\alias{TargetPositionCreationDTO}
\title{TargetPositionCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TargetPositionCreationDTO
}
\description{
TargetPositionCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{target}}{}

\item{\code{position}}{}
}}

\keyword{datasets}
