% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeConfigDTO.r
\docType{data}
\name{ThemeConfigDTO}
\alias{ThemeConfigDTO}
\title{ThemeConfigDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ThemeConfigDTO
}
\description{
ThemeConfigDTO Class
}
\section{Fields}{

\describe{
\item{\code{hasStyle}}{}

\item{\code{fonts}}{}

\item{\code{iconClassesRDF}}{}

\item{\code{componentOverrides}}{}
}}

\keyword{datasets}
