% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenGetDTO.r
\docType{data}
\name{TokenGetDTO}
\alias{TokenGetDTO}
\title{TokenGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenGetDTO
}
\description{
TokenGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{token}}{}
}}

\keyword{datasets}
