% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trait.r
\name{Trait}
\alias{Trait}
\title{Trait Class}
\description{
Trait Class

Trait Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alternativeAbbreviations}}{}

\item{\code{attribute}}{}

\item{\code{description}}{}

\item{\code{entity}}{}

\item{\code{mainAbbreviation}}{}

\item{\code{name}}{}

\item{\code{ontologyReference}}{}

\item{\code{status}}{}

\item{\code{synonyms}}{}

\item{\code{traitDbId}}{}

\item{\code{traitName}}{}

\item{\code{xref}}{}

\item{\code{class}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Trait$new()}}
\item \href{#method-toJSON}{\code{Trait$toJSON()}}
\item \href{#method-fromJSON}{\code{Trait$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{Trait$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{Trait$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Trait$fromJSONString()}}
\item \href{#method-clone}{\code{Trait$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$new(
  alternativeAbbreviations,
  attribute,
  description,
  entity,
  mainAbbreviation,
  name,
  ontologyReference,
  status,
  synonyms,
  traitDbId,
  traitName,
  xref,
  class
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$fromJSON(TraitJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$fromJSONObject(TraitObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$fromJSONString(TraitJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trait$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
