% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitCreationDTO.r
\docType{data}
\name{UnitCreationDTO}
\alias{UnitCreationDTO}
\title{UnitCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnitCreationDTO
}
\description{
UnitCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{symbol}}{}

\item{\code{alternative_symbol}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
