% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitDetailsDTO.r
\docType{data}
\name{UnitDetailsDTO}
\alias{UnitDetailsDTO}
\title{UnitDetailsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnitDetailsDTO
}
\description{
UnitDetailsDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{symbol}}{}

\item{\code{alternative_symbol}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
