% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitGetDTO.r
\docType{data}
\name{UnitGetDTO}
\alias{UnitGetDTO}
\title{UnitGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnitGetDTO
}
\description{
UnitGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
