% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserCreationDTO.r
\docType{data}
\name{UserCreationDTO}
\alias{UserCreationDTO}
\title{UserCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserCreationDTO
}
\description{
UserCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{email}}{}

\item{\code{language}}{}

\item{\code{password}}{}

\item{\code{admin}}{}
}}

\keyword{datasets}
