% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserGetDTO.r
\docType{data}
\name{UserGetDTO}
\alias{UserGetDTO}
\title{UserGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserGetDTO
}
\description{
UserGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{email}}{}

\item{\code{language}}{}

\item{\code{admin}}{}
}}

\keyword{datasets}
