% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableCreationDTO.r
\name{VariableCreationDTO}
\alias{VariableCreationDTO}
\title{VariableCreationDTO Class}
\description{
VariableCreationDTO Class

VariableCreationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{alternative_name}}{}

\item{\code{description}}{}

\item{\code{entity}}{}

\item{\code{characteristic}}{}

\item{\code{trait}}{}

\item{\code{trait_name}}{}

\item{\code{method}}{}

\item{\code{unit}}{}

\item{\code{species}}{}

\item{\code{datatype}}{}

\item{\code{time_interval}}{}

\item{\code{sampling_interval}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariableCreationDTO$new()}}
\item \href{#method-toJSON}{\code{VariableCreationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VariableCreationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VariableCreationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VariableCreationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VariableCreationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VariableCreationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$new(
  uri,
  name,
  alternative_name,
  description,
  entity,
  characteristic,
  trait,
  trait_name,
  method,
  unit,
  species,
  datatype,
  time_interval,
  sampling_interval,
  exact_match,
  close_match,
  broad_match,
  narrow_match
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$fromJSON(VariableCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$fromJSONObject(VariableCreationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$fromJSONString(VariableCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableCreationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
