% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableDatatypeDTO.r
\docType{data}
\name{VariableDatatypeDTO}
\alias{VariableDatatypeDTO}
\title{VariableDatatypeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariableDatatypeDTO
}
\description{
VariableDatatypeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
