% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableGetDTO.r
\docType{data}
\name{VariableGetDTO}
\alias{VariableGetDTO}
\title{VariableGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariableGetDTO
}
\description{
VariableGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{entity}}{}

\item{\code{characteristic}}{}

\item{\code{method}}{}

\item{\code{unit}}{}
}}

\keyword{datasets}
