% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesApi.r
\docType{data}
\name{VariablesApi}
\alias{VariablesApi}
\title{Variables operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariablesApi
}
\description{
opensilexClientToolsR.Variables
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

classic_export_variable_by_ur_is export variable by list of uris


create_characteristic Add a characteristic


create_entity Add an entity


create_interest_entity Add an entity of interest


create_method Add a method


create_unit Add an unit


create_variable Add a variable


create_variables_group Add a variables group


delete_characteristic Delete a characteristic


delete_entity Delete an entity


delete_interest_entity Delete an entity of interest


delete_method Delete a method


delete_unit Delete an unit


delete_variable Delete a variable


delete_variables_group Delete a variables group


details_export_variable_by_ur_is export detailed variable by list of uris


get_characteristic Get a characteristic


get_characteristics_by_ur_is Get detailed characteristics by uris


get_datatypes Get variables datatypes


get_entities_by_ur_is Get detailed entities by uris


get_entity Get an entity


get_interest_entities_by_ur_is Get detailed entities of interest by uris


get_interest_entity Get an entity of interest


get_method Get a method


get_methods_by_ur_is Get detailed methods by uris


get_unit Get an unit


get_units_by_ur_is Get detailed units by uris


get_variable Get a variable


get_variables_by_ur_is Get detailed variables by uris


get_variables_group Get a variables group


get_variables_group_by_ur_is Get variables groups by their URIs


search_characteristics Search characteristics by name


search_entities Search entities by name


search_interest_entity Search entities of interest by name


search_methods Search methods by name


search_units Search units by name


search_variables Search variables


search_variables_details Search detailed variables by name, long name, entity, characteristic, method or unit name


search_variables_groups Search variables groups


update_characteristic Update a characteristic


update_entity Update an entity


update_interest_entity Update an entity of interest


update_method Update a method


update_unit Update an unit


update_variable Update a variable


update_variables_group Update a variables group

}
}

\keyword{datasets}
