% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesGroupCreationDTO.r
\docType{data}
\name{VariablesGroupCreationDTO}
\alias{VariablesGroupCreationDTO}
\title{VariablesGroupCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariablesGroupCreationDTO
}
\description{
VariablesGroupCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{variables}}{}
}}

\keyword{datasets}
