% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesGroupGetDTO.r
\docType{data}
\name{VariablesGroupGetDTO}
\alias{VariablesGroupGetDTO}
\title{VariablesGroupGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariablesGroupGetDTO
}
\description{
VariablesGroupGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{variables}}{}
}}

\keyword{datasets}
