% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableUpdateDTO.r
\docType{data}
\name{VariableUpdateDTO}
\alias{VariableUpdateDTO}
\title{VariableUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariableUpdateDTO
}
\description{
VariableUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{alternative_name}}{}

\item{\code{description}}{}

\item{\code{entity}}{}

\item{\code{entity_of_interest}}{}

\item{\code{characteristic}}{}

\item{\code{trait}}{}

\item{\code{trait_name}}{}

\item{\code{method}}{}

\item{\code{unit}}{}

\item{\code{species}}{}

\item{\code{datatype}}{}

\item{\code{time_interval}}{}

\item{\code{sampling_interval}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
