% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionInfoDTO.r
\name{VersionInfoDTO}
\alias{VersionInfoDTO}
\title{VersionInfoDTO Class}
\description{
VersionInfoDTO Class

VersionInfoDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{}

\item{\code{version}}{}

\item{\code{description}}{}

\item{\code{contact}}{}

\item{\code{license}}{}

\item{\code{modules_version}}{}

\item{\code{external_docs}}{}

\item{\code{api_docs}}{}

\item{\code{git_commit}}{}

\item{\code{github_page}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VersionInfoDTO$new()}}
\item \href{#method-toJSON}{\code{VersionInfoDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VersionInfoDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VersionInfoDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VersionInfoDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VersionInfoDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VersionInfoDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$new(
  title,
  version,
  description,
  contact,
  license,
  modules_version,
  external_docs,
  api_docs,
  git_commit,
  github_page
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$fromJSON(VersionInfoDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$fromJSONObject(VersionInfoDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$fromJSONString(VersionInfoDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
