% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueDataTypeDTO.r
\name{VueDataTypeDTO}
\alias{VueDataTypeDTO}
\title{VueDataTypeDTO Class}
\description{
VueDataTypeDTO Class

VueDataTypeDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{short_uri}}{}

\item{\code{input_component}}{}

\item{\code{view_component}}{}

\item{\code{label_key}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueDataTypeDTO$new()}}
\item \href{#method-toJSON}{\code{VueDataTypeDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VueDataTypeDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VueDataTypeDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VueDataTypeDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VueDataTypeDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VueDataTypeDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$new(uri, short_uri, input_component, view_component, label_key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$fromJSON(VueDataTypeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$fromJSONObject(VueDataTypeDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$fromJSONString(VueDataTypeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueDataTypeDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
