% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueJsApi.r
\docType{data}
\name{VueJsApi}
\alias{VueJsApi}
\title{VueJs operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueJsApi
}
\description{
opensilexClientToolsR.VueJs
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_config Return the current configuration


get_extension Return the front Vue JS extension file to include


get_extension_style Return the front Vue JS extension css file to include


get_theme_config Return the front Vue JS theme configuration


get_theme_css Return the theme css file


get_theme_resource Return the theme requested resource

}
}

\keyword{datasets}
