% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueJsOntologyExtensionApi.r
\docType{data}
\name{VueJsOntologyExtensionApi}
\alias{VueJsOntologyExtensionApi}
\title{VueJsOntologyExtension operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueJsOntologyExtensionApi
}
\description{
opensilexClientToolsR.VueJsOntologyExtension
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_rdf_type Create a custom class


delete_rdf_type Delete a RDF type


get_data_types1 Return literal datatypes definition


get_object_types Return object types definition


get_rdf_type1 Return rdt type model definition with properties


get_rdf_type_properties Return class model properties definitions


get_rdf_types_parameters Return RDF types parameters for Vue.js application


set_rdf_type_properties_order Define properties order


update_rdf_type Update a custom class

}
}

\keyword{datasets}
