% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueObjectTypeDTO.r
\docType{data}
\name{VueObjectTypeDTO}
\alias{VueObjectTypeDTO}
\title{VueObjectTypeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueObjectTypeDTO
}
\description{
VueObjectTypeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{short_uri}}{}

\item{\code{input_component}}{}

\item{\code{input_components_by_property}}{}

\item{\code{view_component}}{}

\item{\code{rdf_type}}{}
}}

\keyword{datasets}
