% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueRDFTypeDTO.r
\docType{data}
\name{VueRDFTypeDTO}
\alias{VueRDFTypeDTO}
\title{VueRDFTypeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueRDFTypeDTO
}
\description{
VueRDFTypeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{parent}}{}

\item{\code{icon}}{}

\item{\code{name_translations}}{}

\item{\code{comment_translations}}{}

\item{\code{is_abstract}}{}

\item{\code{data_properties}}{}

\item{\code{object_properties}}{}

\item{\code{properties_order}}{}
}}

\keyword{datasets}
