% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueRDFTypeParameterDTO.r
\docType{data}
\name{VueRDFTypeParameterDTO}
\alias{VueRDFTypeParameterDTO}
\title{VueRDFTypeParameterDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueRDFTypeParameterDTO
}
\description{
VueRDFTypeParameterDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{icon}}{}

\item{\code{is_abstract}}{}
}}

\keyword{datasets}
