% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueRDFTypePropertyDTO.r
\docType{data}
\name{VueRDFTypePropertyDTO}
\alias{VueRDFTypePropertyDTO}
\title{VueRDFTypePropertyDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueRDFTypePropertyDTO
}
\description{
VueRDFTypePropertyDTO Class
}
\section{Fields}{

\describe{
\item{\code{property}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{inherited}}{}

\item{\code{target_property}}{}

\item{\code{input_component}}{}

\item{\code{input_components_by_property}}{}

\item{\code{view_component}}{}

\item{\code{is_list}}{}

\item{\code{is_required}}{}

\item{\code{is_custom}}{}
}}

\keyword{datasets}
