% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgentModel.r
\docType{data}
\name{AgentModel}
\alias{AgentModel}
\title{AgentModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AgentModel
}
\description{
AgentModel Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{settings}}{}
}}

\keyword{datasets}
