% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationsApi.r
\name{AnnotationsApi}
\alias{AnnotationsApi}
\title{Annotations operations}
\description{
opensilexClientToolsR.Annotations
}
\section{Methods}{

\describe{

count_annotations Count annotations


create_annotation Create an annotation


delete_annotation Delete an annotation


get_annotation Get an annotation


search_annotations Search annotations


search_motivations Search motivations


update_annotation Update an annotation

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotationsApi$new()}}
\item \href{#method-count_annotations}{\code{AnnotationsApi$count_annotations()}}
\item \href{#method-create_annotation}{\code{AnnotationsApi$create_annotation()}}
\item \href{#method-delete_annotation}{\code{AnnotationsApi$delete_annotation()}}
\item \href{#method-get_annotation}{\code{AnnotationsApi$get_annotation()}}
\item \href{#method-search_annotations}{\code{AnnotationsApi$search_annotations()}}
\item \href{#method-search_motivations}{\code{AnnotationsApi$search_motivations()}}
\item \href{#method-update_annotation}{\code{AnnotationsApi$update_annotation()}}
\item \href{#method-clone}{\code{AnnotationsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count_annotations"></a>}}
\if{latex}{\out{\hypertarget{method-count_annotations}{}}}
\subsection{Method \code{count_annotations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$count_annotations(target, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-create_annotation}{}}}
\subsection{Method \code{create_annotation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$create_annotation(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-delete_annotation}{}}}
\subsection{Method \code{delete_annotation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$delete_annotation(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-get_annotation}{}}}
\subsection{Method \code{get_annotation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$get_annotation(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_annotations"></a>}}
\if{latex}{\out{\hypertarget{method-search_annotations}{}}}
\subsection{Method \code{search_annotations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$search_annotations(
  description,
  target,
  motivation,
  author,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_motivations"></a>}}
\if{latex}{\out{\hypertarget{method-search_motivations}{}}}
\subsection{Method \code{search_motivations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$search_motivations(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-update_annotation}{}}}
\subsection{Method \code{update_annotation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$update_annotation(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
