% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiContactInfoDTO.r
\docType{data}
\name{ApiContactInfoDTO}
\alias{ApiContactInfoDTO}
\title{ApiContactInfoDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiContactInfoDTO
}
\description{
ApiContactInfoDTO Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{email}}{}

\item{\code{homepage}}{}
}}

\keyword{datasets}
