% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiExternalDocsDTO.r
\docType{data}
\name{ApiExternalDocsDTO}
\alias{ApiExternalDocsDTO}
\title{ApiExternalDocsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiExternalDocsDTO
}
\description{
ApiExternalDocsDTO Class
}
\section{Fields}{

\describe{
\item{\code{description}}{}

\item{\code{url}}{}
}}

\keyword{datasets}
