% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationApi.r
\docType{data}
\name{AuthenticationApi}
\alias{AuthenticationApi}
\title{Authentication operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AuthenticationApi
}
\description{
opensilexClientToolsR.Authentication
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

authenticate Authenticate a user and return an access token


authenticate_open_id Authenticate a user and return an access token


get_credentials_groups Get list of existing credentials indexed by Swagger @API concepts in the application


logout Logout by discarding a user token


renew_token Send back a new token if the provided one is still valid

}
}

\keyword{datasets}
