% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAPIApi.r
\docType{data}
\name{BRAPIApi}
\alias{BRAPIApi}
\title{BRAPI operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BRAPIApi
}
\description{
opensilexClientToolsR.BRAPI
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_calls Check the available BrAPI calls


get_observation_units List all the observation units measured in the study.


get_observation_variables List all the observation variables measured in the study.


get_observations Get the observations associated to a specific study


get_studies Retrieve studies information


get_studies_search Retrieve studies information


get_study_details Retrieve study details


get_variable_details Retrieve variable details by id


get_variables_list Call to retrieve a list of observationVariables available in the system

}
}

\keyword{datasets}
