% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Call.r
\docType{data}
\name{Call}
\alias{Call}
\title{Call Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Call
}
\description{
Call Class
}
\section{Fields}{

\describe{
\item{\code{call}}{}

\item{\code{dataTypes}}{}

\item{\code{methods}}{}

\item{\code{versions}}{}
}}

\keyword{datasets}
