% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVDuplicateURIError.r
\name{CSVDuplicateURIError}
\alias{CSVDuplicateURIError}
\title{CSVDuplicateURIError Class}
\description{
CSVDuplicateURIError Class

CSVDuplicateURIError Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rowIndex}}{}

\item{\code{colIndex}}{}

\item{\code{header}}{}

\item{\code{value}}{}

\item{\code{previousRow}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSVDuplicateURIError$new()}}
\item \href{#method-toJSON}{\code{CSVDuplicateURIError$toJSON()}}
\item \href{#method-fromJSON}{\code{CSVDuplicateURIError$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{CSVDuplicateURIError$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{CSVDuplicateURIError$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CSVDuplicateURIError$fromJSONString()}}
\item \href{#method-clone}{\code{CSVDuplicateURIError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$new(rowIndex, colIndex, header, value, previousRow)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$fromJSON(CSVDuplicateURIErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$fromJSONObject(CSVDuplicateURIErrorObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$fromJSONString(CSVDuplicateURIErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVDuplicateURIError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
