% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVURINotFoundError.r
\docType{data}
\name{CSVURINotFoundError}
\alias{CSVURINotFoundError}
\title{CSVURINotFoundError Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVURINotFoundError
}
\description{
CSVURINotFoundError Class
}
\section{Fields}{

\describe{
\item{\code{rowIndex}}{}

\item{\code{colIndex}}{}

\item{\code{header}}{}

\item{\code{value}}{}

\item{\code{rdfType}}{}

\item{\code{objectURI}}{}
}}

\keyword{datasets}
