% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVValidationModel.r
\docType{data}
\name{CSVValidationModel}
\alias{CSVValidationModel}
\title{CSVValidationModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVValidationModel
}
\description{
CSVValidationModel Class
}
\section{Fields}{

\describe{
\item{\code{missingHeaders}}{}

\item{\code{invalidHeaderURIs}}{}

\item{\code{datatypeErrors}}{}

\item{\code{uriNotFoundErrors}}{}

\item{\code{invalidURIErrors}}{}

\item{\code{missingRequiredValueErrors}}{}

\item{\code{invalidValueErrors}}{}

\item{\code{alreadyExistingURIErrors}}{}

\item{\code{duplicateURIErrors}}{}
}}

\keyword{datasets}
