% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataApi.r
\name{DataApi}
\alias{DataApi}
\title{Data operations}
\description{
opensilexClientToolsR.Data
}
\section{Methods}{

\describe{

add_list_data Add data


count_data Count data


create_provenance Add a provenance


delete_data Delete data


delete_data_on_search Delete data on criteria


delete_provenance Delete a provenance that doesn&#39;t describe data


export_data Export data


get_data Get data


get_data_file Get a data file


get_data_file_description Get a data file description


get_data_file_descriptions_by_search Search data files


get_datafiles_provenances Get provenances linked to datafiles


get_pictures_thumbnails Get a picture thumbnail


get_provenance Get a provenance


get_provenances_by_ur_is Get a list of provenances by their URIs


get_used_provenances Get provenances linked to data


get_used_variables Get variables linked to data


import_csv_data Import a CSV file for the given provenanceURI


post_data_file Add a data file


post_data_file_paths Describe datafiles and give their relative paths in the configured storage system. In the case of already stored datafiles.


search_data_list Search data


search_provenance Get provenances


update Update data


update_confidence Update confidence index


update_provenance Update a provenance


validate_csv Import a CSV file for the given provenanceURI.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataApi$new()}}
\item \href{#method-add_list_data}{\code{DataApi$add_list_data()}}
\item \href{#method-count_data}{\code{DataApi$count_data()}}
\item \href{#method-create_provenance}{\code{DataApi$create_provenance()}}
\item \href{#method-delete_data}{\code{DataApi$delete_data()}}
\item \href{#method-delete_data_on_search}{\code{DataApi$delete_data_on_search()}}
\item \href{#method-delete_provenance}{\code{DataApi$delete_provenance()}}
\item \href{#method-export_data}{\code{DataApi$export_data()}}
\item \href{#method-get_data}{\code{DataApi$get_data()}}
\item \href{#method-get_data_file}{\code{DataApi$get_data_file()}}
\item \href{#method-get_data_file_description}{\code{DataApi$get_data_file_description()}}
\item \href{#method-get_data_file_descriptions_by_search}{\code{DataApi$get_data_file_descriptions_by_search()}}
\item \href{#method-get_datafiles_provenances}{\code{DataApi$get_datafiles_provenances()}}
\item \href{#method-get_pictures_thumbnails}{\code{DataApi$get_pictures_thumbnails()}}
\item \href{#method-get_provenance}{\code{DataApi$get_provenance()}}
\item \href{#method-get_provenances_by_ur_is}{\code{DataApi$get_provenances_by_ur_is()}}
\item \href{#method-get_used_provenances}{\code{DataApi$get_used_provenances()}}
\item \href{#method-get_used_variables}{\code{DataApi$get_used_variables()}}
\item \href{#method-import_csv_data}{\code{DataApi$import_csv_data()}}
\item \href{#method-post_data_file}{\code{DataApi$post_data_file()}}
\item \href{#method-post_data_file_paths}{\code{DataApi$post_data_file_paths()}}
\item \href{#method-search_data_list}{\code{DataApi$search_data_list()}}
\item \href{#method-search_provenance}{\code{DataApi$search_provenance()}}
\item \href{#method-update}{\code{DataApi$update()}}
\item \href{#method-update_confidence}{\code{DataApi$update_confidence()}}
\item \href{#method-update_provenance}{\code{DataApi$update_provenance()}}
\item \href{#method-validate_csv}{\code{DataApi$validate_csv()}}
\item \href{#method-clone}{\code{DataApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_list_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_list_data}{}}}
\subsection{Method \code{add_list_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$add_list_data(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count_data"></a>}}
\if{latex}{\out{\hypertarget{method-count_data}{}}}
\subsection{Method \code{count_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$count_data(
  start_date,
  end_date,
  timezone,
  experiments,
  targets,
  variables,
  devices,
  min_confidence,
  max_confidence,
  provenances,
  metadata,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_provenance"></a>}}
\if{latex}{\out{\hypertarget{method-create_provenance}{}}}
\subsection{Method \code{create_provenance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$create_provenance(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_data"></a>}}
\if{latex}{\out{\hypertarget{method-delete_data}{}}}
\subsection{Method \code{delete_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$delete_data(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_data_on_search"></a>}}
\if{latex}{\out{\hypertarget{method-delete_data_on_search}{}}}
\subsection{Method \code{delete_data_on_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$delete_data_on_search(experiment, target, variable, provenance, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_provenance"></a>}}
\if{latex}{\out{\hypertarget{method-delete_provenance}{}}}
\subsection{Method \code{delete_provenance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$delete_provenance(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_data"></a>}}
\if{latex}{\out{\hypertarget{method-export_data}{}}}
\subsection{Method \code{export_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$export_data(
  start_date,
  end_date,
  timezone,
  experiments,
  targets,
  variables,
  devices,
  min_confidence,
  max_confidence,
  provenances,
  metadata,
  mode,
  with_raw_data,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_file"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_file}{}}}
\subsection{Method \code{get_data_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_file(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_file_description"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_file_description}{}}}
\subsection{Method \code{get_data_file_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_file_description(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_file_descriptions_by_search"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_file_descriptions_by_search}{}}}
\subsection{Method \code{get_data_file_descriptions_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_file_descriptions_by_search(
  rdf_type,
  start_date,
  end_date,
  timezone,
  experiments,
  targets,
  devices,
  provenances,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datafiles_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-get_datafiles_provenances}{}}}
\subsection{Method \code{get_datafiles_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_datafiles_provenances(experiments, targets, devices, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pictures_thumbnails"></a>}}
\if{latex}{\out{\hypertarget{method-get_pictures_thumbnails}{}}}
\subsection{Method \code{get_pictures_thumbnails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_pictures_thumbnails(uri, scaled_width, scaled_height, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_provenance"></a>}}
\if{latex}{\out{\hypertarget{method-get_provenance}{}}}
\subsection{Method \code{get_provenance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_provenance(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_provenances_by_ur_is"></a>}}
\if{latex}{\out{\hypertarget{method-get_provenances_by_ur_is}{}}}
\subsection{Method \code{get_provenances_by_ur_is()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_provenances_by_ur_is(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_used_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-get_used_provenances}{}}}
\subsection{Method \code{get_used_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_used_provenances(experiments, targets, variables, devices, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_used_variables"></a>}}
\if{latex}{\out{\hypertarget{method-get_used_variables}{}}}
\subsection{Method \code{get_used_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_used_variables(experiments, targets, provenances, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_csv_data"></a>}}
\if{latex}{\out{\hypertarget{method-import_csv_data}{}}}
\subsection{Method \code{import_csv_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$import_csv_data(provenance, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_data_file"></a>}}
\if{latex}{\out{\hypertarget{method-post_data_file}{}}}
\subsection{Method \code{post_data_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$post_data_file(description, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_data_file_paths"></a>}}
\if{latex}{\out{\hypertarget{method-post_data_file_paths}{}}}
\subsection{Method \code{post_data_file_paths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$post_data_file_paths(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-search_data_list}{}}}
\subsection{Method \code{search_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$search_data_list(
  start_date,
  end_date,
  timezone,
  experiments,
  targets,
  variables,
  devices,
  min_confidence,
  max_confidence,
  provenances,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_provenance"></a>}}
\if{latex}{\out{\hypertarget{method-search_provenance}{}}}
\subsection{Method \code{search_provenance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$search_provenance(
  name,
  description,
  activity,
  activity_type,
  agent,
  agent_type,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$update(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_confidence"></a>}}
\if{latex}{\out{\hypertarget{method-update_confidence}{}}}
\subsection{Method \code{update_confidence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$update_confidence(uri, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_provenance"></a>}}
\if{latex}{\out{\hypertarget{method-update_provenance}{}}}
\subsection{Method \code{update_provenance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$update_provenance(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_csv"></a>}}
\if{latex}{\out{\hypertarget{method-validate_csv}{}}}
\subsection{Method \code{validate_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$validate_csv(provenance, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
