% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationDTO.r
\docType{data}
\name{DataCreationDTO}
\alias{DataCreationDTO}
\title{DataCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataCreationDTO
}
\description{
DataCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{date}}{}

\item{\code{timezone}}{}

\item{\code{target}}{}

\item{\code{variable}}{}

\item{\code{value}}{}

\item{\code{confidence}}{}

\item{\code{provenance}}{}

\item{\code{metadata}}{}

\item{\code{raw_data}}{}
}}

\keyword{datasets}
