% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCSVValidationModel.r
\docType{data}
\name{DataCSVValidationModel}
\alias{DataCSVValidationModel}
\title{DataCSVValidationModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataCSVValidationModel
}
\description{
DataCSVValidationModel Class
}
\section{Fields}{

\describe{
\item{\code{missingHeaders}}{}

\item{\code{emptyHeaders}}{}

\item{\code{invalidHeaderURIs}}{}

\item{\code{datatypeErrors}}{}

\item{\code{uriNotFoundErrors}}{}

\item{\code{invalidURIErrors}}{}

\item{\code{missingRequiredValueErrors}}{}

\item{\code{invalidValueErrors}}{}

\item{\code{alreadyExistingURIErrors}}{}

\item{\code{duplicateURIErrors}}{}

\item{\code{invalidObjectErrors}}{}

\item{\code{invalidDateErrors}}{}

\item{\code{invalidDataTypeErrors}}{}

\item{\code{invalidExperimentErrors}}{}

\item{\code{invalidDeviceErrors}}{}

\item{\code{deviceChoiceAmbiguityErrors}}{}

\item{\code{duplicatedDataErrors}}{}

\item{\code{duplicatedObjectErrors}}{}

\item{\code{duplicatedExperimentErrors}}{}

\item{\code{duplicatedDeviceErrors}}{}

\item{\code{headers}}{}

\item{\code{headersLabels}}{}

\item{\code{nbLinesImported}}{}

\item{\code{nbLinesToImport}}{}

\item{\code{validationStep}}{}

\item{\code{insertionStep}}{}

\item{\code{validCSV}}{}

\item{\code{tooLargeDataset}}{}

\item{\code{errorMessage}}{}
}}

\keyword{datasets}
