% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFileGetDTO.r
\docType{data}
\name{DataFileGetDTO}
\alias{DataFileGetDTO}
\title{DataFileGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataFileGetDTO
}
\description{
DataFileGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{date}}{}

\item{\code{timezone}}{}

\item{\code{target}}{}

\item{\code{provenance}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
