% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFilePathCreationDTO.r
\docType{data}
\name{DataFilePathCreationDTO}
\alias{DataFilePathCreationDTO}
\title{DataFilePathCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataFilePathCreationDTO
}
\description{
DataFilePathCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{date}}{}

\item{\code{timezone}}{}

\item{\code{scientific_objects}}{}

\item{\code{provenance}}{}

\item{\code{metadata}}{}

\item{\code{relative_path}}{}
}}

\keyword{datasets}
