% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGetDTO.r
\docType{data}
\name{DataGetDTO}
\alias{DataGetDTO}
\title{DataGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataGetDTO
}
\description{
DataGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{date}}{}

\item{\code{scientific_object}}{}

\item{\code{variable}}{}

\item{\code{value}}{}

\item{\code{confidence}}{}

\item{\code{provenance}}{}

\item{\code{metadata}}{}

\item{\code{raw_data}}{}
}}

\keyword{datasets}
