% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLink.r
\docType{data}
\name{DataLink}
\alias{DataLink}
\title{DataLink Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataLink
}
\description{
DataLink Class
}
\section{Fields}{

\describe{
\item{\code{dataLinkName}}{}

\item{\code{name}}{}

\item{\code{type}}{}

\item{\code{url}}{}
}}

\keyword{datasets}
