% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeviceGetDTO.r
\docType{data}
\name{DeviceGetDTO}
\alias{DeviceGetDTO}
\title{DeviceGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DeviceGetDTO
}
\description{
DeviceGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{brand}}{}

\item{\code{constructor_model}}{}

\item{\code{serial_number}}{}

\item{\code{person_in_charge}}{}

\item{\code{start_up}}{}

\item{\code{removal}}{}

\item{\code{relations}}{}

\item{\code{description}}{}
}}

\keyword{datasets}
