% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevicesApi.r
\name{DevicesApi}
\alias{DevicesApi}
\title{Devices operations}
\description{
opensilexClientToolsR.Devices
}
\section{Methods}{

\describe{

count_device_data Count device data


create_device Create a device


delete_device Delete a device


export_devices export devices


export_list export devices


get_device Get device detail


get_device_by_uris Get devices by uris


get_device_data_files_provenances Get provenances of datafiles linked to this device


get_device_data_provenances Get provenances of data that have been measured on this device


get_device_variables Get variables linked to the device


search_device_data Search device data


search_device_datafiles Search device datafiles descriptions


search_devices Search devices


update_device Update a device

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DevicesApi$new()}}
\item \href{#method-count_device_data}{\code{DevicesApi$count_device_data()}}
\item \href{#method-create_device}{\code{DevicesApi$create_device()}}
\item \href{#method-delete_device}{\code{DevicesApi$delete_device()}}
\item \href{#method-export_devices}{\code{DevicesApi$export_devices()}}
\item \href{#method-export_list}{\code{DevicesApi$export_list()}}
\item \href{#method-get_device}{\code{DevicesApi$get_device()}}
\item \href{#method-get_device_by_uris}{\code{DevicesApi$get_device_by_uris()}}
\item \href{#method-get_device_data_files_provenances}{\code{DevicesApi$get_device_data_files_provenances()}}
\item \href{#method-get_device_data_provenances}{\code{DevicesApi$get_device_data_provenances()}}
\item \href{#method-get_device_variables}{\code{DevicesApi$get_device_variables()}}
\item \href{#method-search_device_data}{\code{DevicesApi$search_device_data()}}
\item \href{#method-search_device_datafiles}{\code{DevicesApi$search_device_datafiles()}}
\item \href{#method-search_devices}{\code{DevicesApi$search_devices()}}
\item \href{#method-update_device}{\code{DevicesApi$update_device()}}
\item \href{#method-clone}{\code{DevicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count_device_data"></a>}}
\if{latex}{\out{\hypertarget{method-count_device_data}{}}}
\subsection{Method \code{count_device_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$count_device_data(
  uri,
  start_date,
  end_date,
  timezone,
  experiment,
  variable,
  min_confidence,
  max_confidence,
  provenance,
  metadata,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_device"></a>}}
\if{latex}{\out{\hypertarget{method-create_device}{}}}
\subsection{Method \code{create_device()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$create_device(body, check_only, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_device"></a>}}
\if{latex}{\out{\hypertarget{method-delete_device}{}}}
\subsection{Method \code{delete_device()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$delete_device(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_devices"></a>}}
\if{latex}{\out{\hypertarget{method-export_devices}{}}}
\subsection{Method \code{export_devices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$export_devices(
  rdf_type,
  include_subtypes,
  name,
  year,
  existence_date,
  brand,
  model,
  serial_number,
  metadata,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_list"></a>}}
\if{latex}{\out{\hypertarget{method-export_list}{}}}
\subsection{Method \code{export_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$export_list(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device"></a>}}
\if{latex}{\out{\hypertarget{method-get_device}{}}}
\subsection{Method \code{get_device()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$get_device(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device_by_uris"></a>}}
\if{latex}{\out{\hypertarget{method-get_device_by_uris}{}}}
\subsection{Method \code{get_device_by_uris()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$get_device_by_uris(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device_data_files_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-get_device_data_files_provenances}{}}}
\subsection{Method \code{get_device_data_files_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$get_device_data_files_provenances(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device_data_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-get_device_data_provenances}{}}}
\subsection{Method \code{get_device_data_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$get_device_data_provenances(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device_variables"></a>}}
\if{latex}{\out{\hypertarget{method-get_device_variables}{}}}
\subsection{Method \code{get_device_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$get_device_variables(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_device_data"></a>}}
\if{latex}{\out{\hypertarget{method-search_device_data}{}}}
\subsection{Method \code{search_device_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$search_device_data(
  uri,
  start_date,
  end_date,
  timezone,
  experiment,
  variable,
  min_confidence,
  max_confidence,
  provenance,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_device_datafiles"></a>}}
\if{latex}{\out{\hypertarget{method-search_device_datafiles}{}}}
\subsection{Method \code{search_device_datafiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$search_device_datafiles(
  uri,
  rdf_type,
  start_date,
  end_date,
  timezone,
  experiment,
  scientific_objects,
  provenances,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_devices"></a>}}
\if{latex}{\out{\hypertarget{method-search_devices}{}}}
\subsection{Method \code{search_devices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$search_devices(
  rdf_type,
  include_subtypes,
  name,
  year,
  existence_date,
  brand,
  model,
  serial_number,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_device"></a>}}
\if{latex}{\out{\hypertarget{method-update_device}{}}}
\subsection{Method \code{update_device()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$update_device(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DevicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
