% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentGetDTO.r
\docType{data}
\name{DocumentGetDTO}
\alias{DocumentGetDTO}
\title{DocumentGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentGetDTO
}
\description{
DocumentGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{identifier}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{title}}{}

\item{\code{date}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{authors}}{}

\item{\code{language}}{}

\item{\code{format}}{}

\item{\code{keywords}}{}

\item{\code{deprecated}}{}
}}

\keyword{datasets}
