% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentsApi.r
\docType{data}
\name{DocumentsApi}
\alias{DocumentsApi}
\title{Documents operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentsApi
}
\description{
opensilexClientToolsR.Documents
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_document Add a document


delete_document Delete a document


get_document_file Get document


get_document_metadata Get document&#39;s description


search_documents Search documents


update_document Update document&#39;s description

}
}

\keyword{datasets}
