% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityGetDTO.r
\docType{data}
\name{EntityGetDTO}
\alias{EntityGetDTO}
\title{EntityGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityGetDTO
}
\description{
EntityGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
