% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventCreationDTO.r
\docType{data}
\name{EventCreationDTO}
\alias{EventCreationDTO}
\title{EventCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventCreationDTO
}
\description{
EventCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{relations}}{}
}}

\keyword{datasets}
