% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventGetDTO.r
\docType{data}
\name{EventGetDTO}
\alias{EventGetDTO}
\title{EventGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventGetDTO
}
\description{
EventGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{author}}{}
}}

\keyword{datasets}
