% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventsApi.r
\name{EventsApi}
\alias{EventsApi}
\title{Events operations}
\description{
opensilexClientToolsR.Events
}
\section{Methods}{

\describe{

count_events Count events


create_events Create a list of event


create_moves Create a list of move event


delete_event Delete an event


delete_move_event Delete a move event


get_event Get an event


get_event_details Get an event with all it&#39;s properties


get_move_event Get a move with all it&#39;s properties


import_event_csv Import a CSV file with one move and one target per line


import_move_csv Import a CSV file with one move and one target per line


search_events Search events


update_event Update an event


update_move_event Update a move event


validate_event_csv Check a CSV file with one move and one target per line


validate_move_csv Check a CSV file with one move and one target per line

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EventsApi$new()}}
\item \href{#method-count_events}{\code{EventsApi$count_events()}}
\item \href{#method-create_events}{\code{EventsApi$create_events()}}
\item \href{#method-create_moves}{\code{EventsApi$create_moves()}}
\item \href{#method-delete_event}{\code{EventsApi$delete_event()}}
\item \href{#method-delete_move_event}{\code{EventsApi$delete_move_event()}}
\item \href{#method-get_event}{\code{EventsApi$get_event()}}
\item \href{#method-get_event_details}{\code{EventsApi$get_event_details()}}
\item \href{#method-get_move_event}{\code{EventsApi$get_move_event()}}
\item \href{#method-import_event_csv}{\code{EventsApi$import_event_csv()}}
\item \href{#method-import_move_csv}{\code{EventsApi$import_move_csv()}}
\item \href{#method-search_events}{\code{EventsApi$search_events()}}
\item \href{#method-update_event}{\code{EventsApi$update_event()}}
\item \href{#method-update_move_event}{\code{EventsApi$update_move_event()}}
\item \href{#method-validate_event_csv}{\code{EventsApi$validate_event_csv()}}
\item \href{#method-validate_move_csv}{\code{EventsApi$validate_move_csv()}}
\item \href{#method-clone}{\code{EventsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count_events"></a>}}
\if{latex}{\out{\hypertarget{method-count_events}{}}}
\subsection{Method \code{count_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$count_events(targets, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_events"></a>}}
\if{latex}{\out{\hypertarget{method-create_events}{}}}
\subsection{Method \code{create_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$create_events(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_moves"></a>}}
\if{latex}{\out{\hypertarget{method-create_moves}{}}}
\subsection{Method \code{create_moves()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$create_moves(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_event"></a>}}
\if{latex}{\out{\hypertarget{method-delete_event}{}}}
\subsection{Method \code{delete_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$delete_event(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_move_event"></a>}}
\if{latex}{\out{\hypertarget{method-delete_move_event}{}}}
\subsection{Method \code{delete_move_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$delete_move_event(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_event"></a>}}
\if{latex}{\out{\hypertarget{method-get_event}{}}}
\subsection{Method \code{get_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$get_event(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_event_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_event_details}{}}}
\subsection{Method \code{get_event_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$get_event_details(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_move_event"></a>}}
\if{latex}{\out{\hypertarget{method-get_move_event}{}}}
\subsection{Method \code{get_move_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$get_move_event(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_event_csv"></a>}}
\if{latex}{\out{\hypertarget{method-import_event_csv}{}}}
\subsection{Method \code{import_event_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$import_event_csv(file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_move_csv"></a>}}
\if{latex}{\out{\hypertarget{method-import_move_csv}{}}}
\subsection{Method \code{import_move_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$import_move_csv(file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_events"></a>}}
\if{latex}{\out{\hypertarget{method-search_events}{}}}
\subsection{Method \code{search_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$search_events(
  rdf_type,
  start,
  end,
  target,
  description,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_event"></a>}}
\if{latex}{\out{\hypertarget{method-update_event}{}}}
\subsection{Method \code{update_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$update_event(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_move_event"></a>}}
\if{latex}{\out{\hypertarget{method-update_move_event}{}}}
\subsection{Method \code{update_move_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$update_move_event(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_event_csv"></a>}}
\if{latex}{\out{\hypertarget{method-validate_event_csv}{}}}
\subsection{Method \code{validate_event_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$validate_event_csv(file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_move_csv"></a>}}
\if{latex}{\out{\hypertarget{method-validate_move_csv}{}}}
\subsection{Method \code{validate_move_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$validate_move_csv(file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
