% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentsApi.r
\name{ExperimentsApi}
\alias{ExperimentsApi}
\title{Experiments operations}
\description{
opensilexClientToolsR.Experiments
}
\section{Methods}{

\describe{

create_experiment Add an experiment


delete_experiment Delete an experiment


export_experiment_data_list export experiment data


get_available_facilities Get facilities available for an experiment


get_available_factors Get factors with their levels associated to an experiment


get_available_species Get species present in an experiment


get_experiment Get an experiment


get_experiments_by_ur_is Get experiments URIs


get_facilities Get facilities involved in an experiment


get_used_variables1 Get variables involved in an experiment


import_csv_data1 Import a CSV file for the given experiment URI and scientific object type.


search_experiment_data_list Search data


search_experiment_provenances Get provenances involved in an experiment


search_experiments Search experiments


update_experiment Update an experiment


validate_csv1 Import a CSV file for the given experiment URI and scientific object type.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExperimentsApi$new()}}
\item \href{#method-create_experiment}{\code{ExperimentsApi$create_experiment()}}
\item \href{#method-delete_experiment}{\code{ExperimentsApi$delete_experiment()}}
\item \href{#method-export_experiment_data_list}{\code{ExperimentsApi$export_experiment_data_list()}}
\item \href{#method-get_available_facilities}{\code{ExperimentsApi$get_available_facilities()}}
\item \href{#method-get_available_factors}{\code{ExperimentsApi$get_available_factors()}}
\item \href{#method-get_available_species}{\code{ExperimentsApi$get_available_species()}}
\item \href{#method-get_experiment}{\code{ExperimentsApi$get_experiment()}}
\item \href{#method-get_experiments_by_ur_is}{\code{ExperimentsApi$get_experiments_by_ur_is()}}
\item \href{#method-get_facilities}{\code{ExperimentsApi$get_facilities()}}
\item \href{#method-get_used_variables1}{\code{ExperimentsApi$get_used_variables1()}}
\item \href{#method-import_csv_data1}{\code{ExperimentsApi$import_csv_data1()}}
\item \href{#method-search_experiment_data_list}{\code{ExperimentsApi$search_experiment_data_list()}}
\item \href{#method-search_experiment_provenances}{\code{ExperimentsApi$search_experiment_provenances()}}
\item \href{#method-search_experiments}{\code{ExperimentsApi$search_experiments()}}
\item \href{#method-update_experiment}{\code{ExperimentsApi$update_experiment()}}
\item \href{#method-validate_csv1}{\code{ExperimentsApi$validate_csv1()}}
\item \href{#method-clone}{\code{ExperimentsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_experiment"></a>}}
\if{latex}{\out{\hypertarget{method-create_experiment}{}}}
\subsection{Method \code{create_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$create_experiment(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_experiment"></a>}}
\if{latex}{\out{\hypertarget{method-delete_experiment}{}}}
\subsection{Method \code{delete_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$delete_experiment(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_experiment_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-export_experiment_data_list}{}}}
\subsection{Method \code{export_experiment_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$export_experiment_data_list(
  uri,
  start_date,
  end_date,
  timezone,
  scientific_objects,
  variables,
  min_confidence,
  max_confidence,
  provenance,
  metadata,
  mode,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_available_facilities"></a>}}
\if{latex}{\out{\hypertarget{method-get_available_facilities}{}}}
\subsection{Method \code{get_available_facilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_available_facilities(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_available_factors"></a>}}
\if{latex}{\out{\hypertarget{method-get_available_factors}{}}}
\subsection{Method \code{get_available_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_available_factors(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_available_species"></a>}}
\if{latex}{\out{\hypertarget{method-get_available_species}{}}}
\subsection{Method \code{get_available_species()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_available_species(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiment"></a>}}
\if{latex}{\out{\hypertarget{method-get_experiment}{}}}
\subsection{Method \code{get_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_experiment(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiments_by_ur_is"></a>}}
\if{latex}{\out{\hypertarget{method-get_experiments_by_ur_is}{}}}
\subsection{Method \code{get_experiments_by_ur_is()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_experiments_by_ur_is(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_facilities"></a>}}
\if{latex}{\out{\hypertarget{method-get_facilities}{}}}
\subsection{Method \code{get_facilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_facilities(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_used_variables1"></a>}}
\if{latex}{\out{\hypertarget{method-get_used_variables1}{}}}
\subsection{Method \code{get_used_variables1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_used_variables1(uri, scientific_objects, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_csv_data1"></a>}}
\if{latex}{\out{\hypertarget{method-import_csv_data1}{}}}
\subsection{Method \code{import_csv_data1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$import_csv_data1(uri, provenance, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_experiment_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-search_experiment_data_list}{}}}
\subsection{Method \code{search_experiment_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$search_experiment_data_list(
  uri,
  start_date,
  end_date,
  timezone,
  scientific_objects,
  variables,
  min_confidence,
  max_confidence,
  provenances,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_experiment_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-search_experiment_provenances}{}}}
\subsection{Method \code{search_experiment_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$search_experiment_provenances(
  uri,
  name,
  description,
  activity,
  activity_type,
  agent,
  agent_type,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_experiments"></a>}}
\if{latex}{\out{\hypertarget{method-search_experiments}{}}}
\subsection{Method \code{search_experiments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$search_experiments(
  name,
  year,
  is_ended,
  species,
  factors,
  projects,
  is_public,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_experiment"></a>}}
\if{latex}{\out{\hypertarget{method-update_experiment}{}}}
\subsection{Method \code{update_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$update_experiment(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_csv1"></a>}}
\if{latex}{\out{\hypertarget{method-validate_csv1}{}}}
\subsection{Method \code{validate_csv1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$validate_csv1(uri, provenance, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
