% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDetailsGetDTO.r
\name{FactorDetailsGetDTO}
\alias{FactorDetailsGetDTO}
\title{FactorDetailsGetDTO Class}
\description{
FactorDetailsGetDTO Class

FactorDetailsGetDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{category}}{}

\item{\code{description}}{}

\item{\code{levels}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}

\item{\code{experiment}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FactorDetailsGetDTO$new()}}
\item \href{#method-toJSON}{\code{FactorDetailsGetDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{FactorDetailsGetDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{FactorDetailsGetDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{FactorDetailsGetDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FactorDetailsGetDTO$fromJSONString()}}
\item \href{#method-clone}{\code{FactorDetailsGetDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$new(
  uri,
  name,
  category,
  description,
  levels,
  exact_match,
  close_match,
  broad_match,
  narrow_match,
  experiment
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$fromJSON(FactorDetailsGetDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$fromJSONObject(FactorDetailsGetDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$fromJSONString(FactorDetailsGetDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorDetailsGetDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
