% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorLevelCreationDTO.r
\docType{data}
\name{FactorLevelCreationDTO}
\alias{FactorLevelCreationDTO}
\title{FactorLevelCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorLevelCreationDTO
}
\description{
FactorLevelCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}
}}

\keyword{datasets}
