% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorLevelGetDTO.r
\docType{data}
\name{FactorLevelGetDTO}
\alias{FactorLevelGetDTO}
\title{FactorLevelGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorLevelGetDTO
}
\description{
FactorLevelGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}
}}

\keyword{datasets}
