% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorsApi.r
\docType{data}
\name{FactorsApi}
\alias{FactorsApi}
\title{Factors operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorsApi
}
\description{
opensilexClientToolsR.Factors
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_factor Create a factor


delete_factor Delete a factor


delete_factor_level Delete a factor level


get_factor_associated_experiments Get factor associated experiments


get_factor_by_uri Get a factor


get_factor_level Get a factor level


get_factor_level_detail Get a factor level


get_factor_levels Get factor levels


get_factors_by_ur_is Get a list of factors by their URIs


search_categories Search categories


search_factor_levels Search factors levels


search_factors Search factors


update_factor Update a factor

}
}

\keyword{datasets}
