% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorUpdateDTO.r
\docType{data}
\name{FactorUpdateDTO}
\alias{FactorUpdateDTO}
\title{FactorUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorUpdateDTO
}
\description{
FactorUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{category}}{}

\item{\code{description}}{}

\item{\code{levels}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
