% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormUpdateDTO.r
\docType{data}
\name{FormUpdateDTO}
\alias{FormUpdateDTO}
\title{FormUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FormUpdateDTO
}
\description{
FormUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{timezone}}{}

\item{\code{uri}}{}

\item{\code{modified_date}}{}

\item{\code{form_data}}{}

\item{\code{commit_address}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
