% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontConfigDTO.r
\docType{data}
\name{FrontConfigDTO}
\alias{FrontConfigDTO}
\title{FrontConfigDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FrontConfigDTO
}
\description{
FrontConfigDTO Class
}
\section{Fields}{

\describe{
\item{\code{pathPrefix}}{}

\item{\code{homeComponent}}{}

\item{\code{notFoundComponent}}{}

\item{\code{headerComponent}}{}

\item{\code{loginComponent}}{}

\item{\code{menuComponent}}{}

\item{\code{footerComponent}}{}

\item{\code{menu}}{}

\item{\code{routes}}{}

\item{\code{themeModule}}{}

\item{\code{themeName}}{}

\item{\code{openIDAuthenticationURI}}{}

\item{\code{openIDConnectionTitle}}{}

\item{\code{activateResetPassword}}{}
}}

\keyword{datasets}
