% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmApi.r
\name{GermplasmApi}
\alias{GermplasmApi}
\title{Germplasm operations}
\description{
opensilexClientToolsR.Germplasm
}
\section{Methods}{

\describe{

create_germplasm Add a germplasm


delete_germplasm Delete a germplasm


export_germplasm export germplasm


export_germplasm_by_ur_is export germplasm by list of uris


get_germplasm Get a germplasm


get_germplasm_experiments Get experiments where a germplasm has been used


get_germplasms_by_uri Get a list of germplasms by their URIs


search_germplasm Search germplasm


update_germplasm Update a germplasm

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GermplasmApi$new()}}
\item \href{#method-create_germplasm}{\code{GermplasmApi$create_germplasm()}}
\item \href{#method-delete_germplasm}{\code{GermplasmApi$delete_germplasm()}}
\item \href{#method-export_germplasm}{\code{GermplasmApi$export_germplasm()}}
\item \href{#method-export_germplasm_by_ur_is}{\code{GermplasmApi$export_germplasm_by_ur_is()}}
\item \href{#method-get_germplasm}{\code{GermplasmApi$get_germplasm()}}
\item \href{#method-get_germplasm_experiments}{\code{GermplasmApi$get_germplasm_experiments()}}
\item \href{#method-get_germplasms_by_uri}{\code{GermplasmApi$get_germplasms_by_uri()}}
\item \href{#method-search_germplasm}{\code{GermplasmApi$search_germplasm()}}
\item \href{#method-update_germplasm}{\code{GermplasmApi$update_germplasm()}}
\item \href{#method-clone}{\code{GermplasmApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_germplasm"></a>}}
\if{latex}{\out{\hypertarget{method-create_germplasm}{}}}
\subsection{Method \code{create_germplasm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$create_germplasm(body, check_only, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_germplasm"></a>}}
\if{latex}{\out{\hypertarget{method-delete_germplasm}{}}}
\subsection{Method \code{delete_germplasm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$delete_germplasm(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_germplasm"></a>}}
\if{latex}{\out{\hypertarget{method-export_germplasm}{}}}
\subsection{Method \code{export_germplasm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$export_germplasm(
  uri,
  rdf_type,
  name,
  code,
  production_year,
  species,
  variety,
  accession,
  institute,
  experiment,
  metadata,
  order_by,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_germplasm_by_ur_is"></a>}}
\if{latex}{\out{\hypertarget{method-export_germplasm_by_ur_is}{}}}
\subsection{Method \code{export_germplasm_by_ur_is()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$export_germplasm_by_ur_is(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_germplasm"></a>}}
\if{latex}{\out{\hypertarget{method-get_germplasm}{}}}
\subsection{Method \code{get_germplasm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$get_germplasm(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_germplasm_experiments"></a>}}
\if{latex}{\out{\hypertarget{method-get_germplasm_experiments}{}}}
\subsection{Method \code{get_germplasm_experiments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$get_germplasm_experiments(
  uri,
  name,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_germplasms_by_uri"></a>}}
\if{latex}{\out{\hypertarget{method-get_germplasms_by_uri}{}}}
\subsection{Method \code{get_germplasms_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$get_germplasms_by_uri(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_germplasm"></a>}}
\if{latex}{\out{\hypertarget{method-search_germplasm}{}}}
\subsection{Method \code{search_germplasm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$search_germplasm(
  uri,
  rdf_type,
  name,
  code,
  production_year,
  species,
  variety,
  accession,
  institute,
  experiment,
  metadata,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_germplasm"></a>}}
\if{latex}{\out{\hypertarget{method-update_germplasm}{}}}
\subsection{Method \code{update_germplasm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$update_germplasm(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
