% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmDTO.r
\name{GermplasmDTO}
\alias{GermplasmDTO}
\title{GermplasmDTO Class}
\description{
GermplasmDTO Class

GermplasmDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessionNumber}}{}

\item{\code{acquisitionDate}}{}

\item{\code{additionalInfo}}{}

\item{\code{biologicalStatusOfAccessionCode}}{}

\item{\code{biologicalStatusOfAccessionDescription}}{}

\item{\code{breedingMethodDbId}}{}

\item{\code{collection}}{}

\item{\code{commonCropName}}{}

\item{\code{countryOfOriginCode}}{}

\item{\code{defaultDisplayName}}{}

\item{\code{documentationURL}}{}

\item{\code{donors}}{}

\item{\code{externalReferences}}{}

\item{\code{genus}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{germplasmOrigin}}{}

\item{\code{germplasmPreprocessing}}{}

\item{\code{instituteCode}}{}

\item{\code{instituteName}}{}

\item{\code{pedigree}}{}

\item{\code{seedSource}}{}

\item{\code{seedSourceDescription}}{}

\item{\code{species}}{}

\item{\code{speciesAuthority}}{}

\item{\code{storageTypes}}{}

\item{\code{subtaxa}}{}

\item{\code{subtaxaAuthority}}{}

\item{\code{synonyms}}{}

\item{\code{taxonIds}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GermplasmDTO$new()}}
\item \href{#method-toJSON}{\code{GermplasmDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{GermplasmDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{GermplasmDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{GermplasmDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GermplasmDTO$fromJSONString()}}
\item \href{#method-clone}{\code{GermplasmDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$new(
  accessionNumber,
  acquisitionDate,
  additionalInfo,
  biologicalStatusOfAccessionCode,
  biologicalStatusOfAccessionDescription,
  breedingMethodDbId,
  collection,
  commonCropName,
  countryOfOriginCode,
  defaultDisplayName,
  documentationURL,
  donors,
  externalReferences,
  genus,
  germplasmDbId,
  germplasmName,
  germplasmOrigin,
  germplasmPreprocessing,
  instituteCode,
  instituteName,
  pedigree,
  seedSource,
  seedSourceDescription,
  species,
  speciesAuthority,
  storageTypes,
  subtaxa,
  subtaxaAuthority,
  synonyms,
  taxonIds
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$fromJSON(GermplasmDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$fromJSONObject(GermplasmDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$fromJSONString(GermplasmDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
