% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmGetAllDTO.r
\docType{data}
\name{GermplasmGetAllDTO}
\alias{GermplasmGetAllDTO}
\title{GermplasmGetAllDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmGetAllDTO
}
\description{
GermplasmGetAllDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{species}}{}

\item{\code{species_name}}{}
}}

\keyword{datasets}
