% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmGetSingleDTO.r
\docType{data}
\name{GermplasmGetSingleDTO}
\alias{GermplasmGetSingleDTO}
\title{GermplasmGetSingleDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmGetSingleDTO
}
\description{
GermplasmGetSingleDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{synonyms}}{}

\item{\code{code}}{}

\item{\code{production_year}}{}

\item{\code{description}}{}

\item{\code{species}}{}

\item{\code{species_name}}{}

\item{\code{variety}}{}

\item{\code{variety_name}}{}

\item{\code{accession}}{}

\item{\code{accession_name}}{}

\item{\code{institute}}{}

\item{\code{website}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
