% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmUpdateDTO.r
\docType{data}
\name{GermplasmUpdateDTO}
\alias{GermplasmUpdateDTO}
\title{GermplasmUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmUpdateDTO
}
\description{
GermplasmUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{synonyms}}{}

\item{\code{code}}{}

\item{\code{production_year}}{}

\item{\code{description}}{}

\item{\code{species}}{}

\item{\code{variety}}{}

\item{\code{accession}}{}

\item{\code{institute}}{}

\item{\code{website}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
