% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupCreationDTO.r
\docType{data}
\name{GroupCreationDTO}
\alias{GroupCreationDTO}
\title{GroupCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupCreationDTO
}
\description{
GroupCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{user_profiles}}{}
}}

\keyword{datasets}
