% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupUserProfileDTO.r
\docType{data}
\name{GroupUserProfileDTO}
\alias{GroupUserProfileDTO}
\title{GroupUserProfileDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupUserProfileDTO
}
\description{
GroupUserProfileDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{profile_uri}}{}

\item{\code{profile_name}}{}

\item{\code{user_uri}}{}

\item{\code{user_name}}{}
}}

\keyword{datasets}
