% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureCreationDTO.r
\docType{data}
\name{InfrastructureCreationDTO}
\alias{InfrastructureCreationDTO}
\title{InfrastructureCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureCreationDTO
}
\description{
InfrastructureCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{parents}}{}

\item{\code{groups}}{}

\item{\code{facilities}}{}
}}

\keyword{datasets}
