% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureFacilityUpdateDTO.r
\docType{data}
\name{InfrastructureFacilityUpdateDTO}
\alias{InfrastructureFacilityUpdateDTO}
\title{InfrastructureFacilityUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructureFacilityUpdateDTO
}
\description{
InfrastructureFacilityUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{organisation}}{}
}}

\keyword{datasets}
