% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MenuItemDTO.r
\docType{data}
\name{MenuItemDTO}
\alias{MenuItemDTO}
\title{MenuItemDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MenuItemDTO
}
\description{
MenuItemDTO Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{children}}{}

\item{\code{route}}{}
}}

\keyword{datasets}
