% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDTO.r
\docType{data}
\name{MetadataDTO}
\alias{MetadataDTO}
\title{MetadataDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MetadataDTO
}
\description{
MetadataDTO Class
}
\section{Fields}{

\describe{
\item{\code{pagination}}{}

\item{\code{status}}{}

\item{\code{datafiles}}{}
}}

\keyword{datasets}
