% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodGetDTO.r
\docType{data}
\name{MethodGetDTO}
\alias{MethodGetDTO}
\title{MethodGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MethodGetDTO
}
\description{
MethodGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
