% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MobileApi.r
\docType{data}
\name{MobileApi}
\alias{MobileApi}
\title{Mobile operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MobileApi
}
\description{
opensilexClientToolsR.Mobile
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_form Add a form


delete_form Delete form


search_form_list Search forms


update1 Update form

}
}

\keyword{datasets}
