% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveCreationDTO.r
\docType{data}
\name{MoveCreationDTO}
\alias{MoveCreationDTO}
\title{MoveCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MoveCreationDTO
}
\description{
MoveCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{relations}}{}

\item{\code{from}}{}

\item{\code{to}}{}

\item{\code{targets_positions}}{}
}}

\keyword{datasets}
