% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveDetailsDTO.r
\docType{data}
\name{MoveDetailsDTO}
\alias{MoveDetailsDTO}
\title{MoveDetailsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MoveDetailsDTO
}
\description{
MoveDetailsDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{author}}{}

\item{\code{relations}}{}

\item{\code{from}}{}

\item{\code{to}}{}

\item{\code{targets_positions}}{}
}}

\keyword{datasets}
