% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveUpdateDTO.r
\docType{data}
\name{MoveUpdateDTO}
\alias{MoveUpdateDTO}
\title{MoveUpdateDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MoveUpdateDTO
}
\description{
MoveUpdateDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{start}}{}

\item{\code{end}}{}

\item{\code{is_instant}}{}

\item{\code{description}}{}

\item{\code{targets}}{}

\item{\code{relations}}{}

\item{\code{from}}{}

\item{\code{to}}{}

\item{\code{targets_positions}}{}
}}

\keyword{datasets}
