% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationSummary.r
\docType{data}
\name{ObservationSummary}
\alias{ObservationSummary}
\title{ObservationSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ObservationSummary
}
\description{
ObservationSummary Class
}
\section{Fields}{

\describe{
\item{\code{collector}}{}

\item{\code{observationDbId}}{}

\item{\code{observationTimeStamp}}{}

\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{season}}{}
}}

\keyword{datasets}
