% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationUnitDTO.r
\docType{data}
\name{ObservationUnitDTO}
\alias{ObservationUnitDTO}
\title{ObservationUnitDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ObservationUnitDTO
}
\description{
ObservationUnitDTO Class
}
\section{Fields}{

\describe{
\item{\code{blockNumber}}{}

\item{\code{entryNumber}}{}

\item{\code{entryType}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{observationLevel}}{}

\item{\code{observationLevels}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitXref}}{}

\item{\code{observations}}{}

\item{\code{plantNumber}}{}

\item{\code{plotNumber}}{}

\item{\code{positionCoordinateX}}{}

\item{\code{positionCoordinateXType}}{}

\item{\code{positionCoordinateY}}{}

\item{\code{positionCoordinateYType}}{}

\item{\code{programName}}{}

\item{\code{replicate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyLocation}}{}

\item{\code{studyLocationDbId}}{}

\item{\code{studyName}}{}

\item{\code{treatments}}{}
}}

\keyword{datasets}
