% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyApi.r
\docType{data}
\name{OntologyApi}
\alias{OntologyApi}
\title{Ontology operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OntologyApi
}
\description{
opensilexClientToolsR.Ontology
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

add_class_property_restriction Add a rdf type property restriction


check_ur_is_types Check the given rdf-types on the given uris


create_property Create a RDF property


delete_class_property_restriction Delete a rdf type property restriction


delete_property Delete a property


get_classes Return classes models definitions with properties for a list of rdt types


get_data_properties Search data properties tree


get_object_properties Search object properties tree


get_properties Search properties tree


get_property Return property model definition detail


get_rdf_type Return class model definition with properties


get_sub_classes_of Search sub-classes tree of an RDF class


get_uri_label Return associated rdfs:label of an uri if exists


update_class_property_restriction Update a rdf type property restriction


update_property Update a RDF property

}
}

\keyword{datasets}
