% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganisationsApi.r
\docType{data}
\name{OrganisationsApi}
\alias{OrganisationsApi}
\title{Organisations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OrganisationsApi
}
\description{
opensilexClientToolsR.Organisations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_infrastructure Create an organisation


create_infrastructure_facility Create a facility


create_infrastructure_team Create a team


delete_infrastructure Delete an organisation


delete_infrastructure_facility Delete a facility


delete_infrastructure_team Delete a team


get_all_facilities Get all facilities


get_facilities_by_uri Get facilities by their URIs


get_infrastructure Get an organisation 


get_infrastructure_facility Get a facility


get_infrastructure_team Get a team


search_infrastructure_facilities Search facilities


search_infrastructures_tree Search organisations


update_infrastructure Update an organisation


update_infrastructure_facility Update a facility


update_infrastructure_team Update a team

}
}

\keyword{datasets}
