% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWLClassPropertyRestrictionDTO.r
\name{OWLClassPropertyRestrictionDTO}
\alias{OWLClassPropertyRestrictionDTO}
\title{OWLClassPropertyRestrictionDTO Class}
\description{
OWLClassPropertyRestrictionDTO Class

OWLClassPropertyRestrictionDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{property}}{}

\item{\code{required}}{}

\item{\code{list}}{}

\item{\code{rdf_type}}{}

\item{\code{domain}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OWLClassPropertyRestrictionDTO$new()}}
\item \href{#method-toJSON}{\code{OWLClassPropertyRestrictionDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{OWLClassPropertyRestrictionDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{OWLClassPropertyRestrictionDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{OWLClassPropertyRestrictionDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OWLClassPropertyRestrictionDTO$fromJSONString()}}
\item \href{#method-clone}{\code{OWLClassPropertyRestrictionDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$new(property, required, list, rdf_type, domain)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$fromJSON(OWLClassPropertyRestrictionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$fromJSONObject(
  OWLClassPropertyRestrictionDTOObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$fromJSONString(
  OWLClassPropertyRestrictionDTOJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWLClassPropertyRestrictionDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
