% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionsApi.r
\docType{data}
\name{PositionsApi}
\alias{PositionsApi}
\title{Positions operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PositionsApi
}
\description{
opensilexClientToolsR.Positions
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_position Get the position of an object


search_position_history Search history of position of an object

}
}

\keyword{datasets}
