% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileCreationDTO.r
\docType{data}
\name{ProfileCreationDTO}
\alias{ProfileCreationDTO}
\title{ProfileCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProfileCreationDTO
}
\description{
ProfileCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{credentials}}{}
}}

\keyword{datasets}
