% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectGetDTO.r
\docType{data}
\name{ProjectGetDTO}
\alias{ProjectGetDTO}
\title{ProjectGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectGetDTO
}
\description{
ProjectGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{start_date}}{}

\item{\code{end_date}}{}

\item{\code{description}}{}

\item{\code{objective}}{}

\item{\code{financial_funding}}{}

\item{\code{website}}{}

\item{\code{related_projects}}{}

\item{\code{coordinators}}{}

\item{\code{scientific_contacts}}{}

\item{\code{administrative_contacts}}{}

\item{\code{experiments}}{}
}}

\keyword{datasets}
