% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectsApi.r
\docType{data}
\name{ProjectsApi}
\alias{ProjectsApi}
\title{Projects operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectsApi
}
\description{
opensilexClientToolsR.Projects
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_project Add a project


delete_project Delete a project


get_project Get a project


get_projects_by_uri Get projects by their URIs


search_projects Search projects


update_project Update a project

}
}

\keyword{datasets}
