% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDFObjectRelationDTO.r
\docType{data}
\name{RDFObjectRelationDTO}
\alias{RDFObjectRelationDTO}
\title{RDFObjectRelationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RDFObjectRelationDTO
}
\description{
RDFObjectRelationDTO Class
}
\section{Fields}{

\describe{
\item{\code{property}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
