% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDFPropertyDTO.r
\name{RDFPropertyDTO}
\alias{RDFPropertyDTO}
\title{RDFPropertyDTO Class}
\description{
RDFPropertyDTO Class

RDFPropertyDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{domain}}{}

\item{\code{range}}{}

\item{\code{parent}}{}

\item{\code{rdf_type}}{}

\item{\code{name_translations}}{}

\item{\code{comment_translations}}{}

\item{\code{domain_rdf_type}}{}

\item{\code{range_label}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RDFPropertyDTO$new()}}
\item \href{#method-toJSON}{\code{RDFPropertyDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{RDFPropertyDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{RDFPropertyDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{RDFPropertyDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RDFPropertyDTO$fromJSONString()}}
\item \href{#method-clone}{\code{RDFPropertyDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$new(
  uri,
  name,
  comment,
  domain,
  range,
  parent,
  rdf_type,
  name_translations,
  comment_translations,
  domain_rdf_type,
  range_label
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$fromJSON(RDFPropertyDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$fromJSONObject(RDFPropertyDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$fromJSONString(RDFPropertyDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDFPropertyDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
