% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDFTypeDTO.r
\docType{data}
\name{RDFTypeDTO}
\alias{RDFTypeDTO}
\title{RDFTypeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RDFTypeDTO
}
\description{
RDFTypeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{parent}}{}
}}

\keyword{datasets}
