% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectCreationDTO.r
\docType{data}
\name{ScientificObjectCreationDTO}
\alias{ScientificObjectCreationDTO}
\title{ScientificObjectCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectCreationDTO
}
\description{
ScientificObjectCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{name}}{}

\item{\code{experiment}}{}

\item{\code{relations}}{}

\item{\code{geometry}}{}
}}

\keyword{datasets}
