% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectDetailByExperimentsDTO.r
\docType{data}
\name{ScientificObjectDetailByExperimentsDTO}
\alias{ScientificObjectDetailByExperimentsDTO}
\title{ScientificObjectDetailByExperimentsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectDetailByExperimentsDTO
}
\description{
ScientificObjectDetailByExperimentsDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{name}}{}

\item{\code{parent}}{}

\item{\code{parent_name}}{}

\item{\code{experiment}}{}

\item{\code{experiment_name}}{}

\item{\code{factor_level}}{}

\item{\code{relations}}{}

\item{\code{geometry}}{}
}}

\keyword{datasets}
