% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectNodeDTO.r
\name{ScientificObjectNodeDTO}
\alias{ScientificObjectNodeDTO}
\title{ScientificObjectNodeDTO Class}
\description{
ScientificObjectNodeDTO Class

ScientificObjectNodeDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{geometry}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{creation_date}}{}

\item{\code{destruction_date}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectNodeDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectNodeDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectNodeDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectNodeDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectNodeDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectNodeDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectNodeDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$new(
  uri,
  name,
  geometry,
  rdf_type,
  rdf_type_name,
  creation_date,
  destruction_date
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$fromJSON(ScientificObjectNodeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$fromJSONObject(ScientificObjectNodeDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$fromJSONString(ScientificObjectNodeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectNodeDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
