% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectNodeWithChildrenDTO.r
\docType{data}
\name{ScientificObjectNodeWithChildrenDTO}
\alias{ScientificObjectNodeWithChildrenDTO}
\title{ScientificObjectNodeWithChildrenDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectNodeWithChildrenDTO
}
\description{
ScientificObjectNodeWithChildrenDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{geometry}}{}

\item{\code{rdf_type}}{}

\item{\code{rdf_type_name}}{}

\item{\code{creation_date}}{}

\item{\code{destruction_date}}{}

\item{\code{child_count}}{}
}}

\keyword{datasets}
