% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectsApi.r
\name{ScientificObjectsApi}
\alias{ScientificObjectsApi}
\title{ScientificObjects operations}
\description{
opensilexClientToolsR.ScientificObjects
}
\section{Methods}{

\describe{

create_scientific_object Create a scientific object for the given experiment


delete_scientific_object Delete a scientific object


export_csv Export a given list of scientific object URIs to csv data file


get_scientific_object_data_files_provenances Get provenances of datafiles linked to this scientific object


get_scientific_object_data_provenances Get provenances of data that have been measured on this scientific object


get_scientific_object_detail Get scientific object detail


get_scientific_object_detail_by_experiments Get scientific object detail for each experiments, a null value for experiment in response means a properties defined outside of any experiment (shared object).


get_scientific_object_variables Get variables measured on this scientific object


get_scientific_objects_children Get list of scientific object children


get_scientific_objects_list_by_uris Get scientific objet list of a given experiment URI


get_used_types get used scientific object types


import_csv Import a CSV file for the given experiment URI and scientific object type.


search_scientific_objects Search list of scientific objects


search_scientific_objects_with_geometry_list_by_uris Get scientific objet list with geometry of a given experiment URI


update_scientific_object Update a scientific object for the given experiment


validate_csv2 Validate a CSV file for the given experiment URI and scientific object type.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectsApi$new()}}
\item \href{#method-create_scientific_object}{\code{ScientificObjectsApi$create_scientific_object()}}
\item \href{#method-delete_scientific_object}{\code{ScientificObjectsApi$delete_scientific_object()}}
\item \href{#method-export_csv}{\code{ScientificObjectsApi$export_csv()}}
\item \href{#method-get_scientific_object_data_files_provenances}{\code{ScientificObjectsApi$get_scientific_object_data_files_provenances()}}
\item \href{#method-get_scientific_object_data_provenances}{\code{ScientificObjectsApi$get_scientific_object_data_provenances()}}
\item \href{#method-get_scientific_object_detail}{\code{ScientificObjectsApi$get_scientific_object_detail()}}
\item \href{#method-get_scientific_object_detail_by_experiments}{\code{ScientificObjectsApi$get_scientific_object_detail_by_experiments()}}
\item \href{#method-get_scientific_object_variables}{\code{ScientificObjectsApi$get_scientific_object_variables()}}
\item \href{#method-get_scientific_objects_children}{\code{ScientificObjectsApi$get_scientific_objects_children()}}
\item \href{#method-get_scientific_objects_list_by_uris}{\code{ScientificObjectsApi$get_scientific_objects_list_by_uris()}}
\item \href{#method-get_used_types}{\code{ScientificObjectsApi$get_used_types()}}
\item \href{#method-import_csv}{\code{ScientificObjectsApi$import_csv()}}
\item \href{#method-search_scientific_objects}{\code{ScientificObjectsApi$search_scientific_objects()}}
\item \href{#method-search_scientific_objects_with_geometry_list_by_uris}{\code{ScientificObjectsApi$search_scientific_objects_with_geometry_list_by_uris()}}
\item \href{#method-update_scientific_object}{\code{ScientificObjectsApi$update_scientific_object()}}
\item \href{#method-validate_csv2}{\code{ScientificObjectsApi$validate_csv2()}}
\item \href{#method-clone}{\code{ScientificObjectsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_scientific_object"></a>}}
\if{latex}{\out{\hypertarget{method-create_scientific_object}{}}}
\subsection{Method \code{create_scientific_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$create_scientific_object(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_scientific_object"></a>}}
\if{latex}{\out{\hypertarget{method-delete_scientific_object}{}}}
\subsection{Method \code{delete_scientific_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$delete_scientific_object(uri, experiment, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_csv"></a>}}
\if{latex}{\out{\hypertarget{method-export_csv}{}}}
\subsection{Method \code{export_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$export_csv(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_object_data_files_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_object_data_files_provenances}{}}}
\subsection{Method \code{get_scientific_object_data_files_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_object_data_files_provenances(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_object_data_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_object_data_provenances}{}}}
\subsection{Method \code{get_scientific_object_data_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_object_data_provenances(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_object_detail"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_object_detail}{}}}
\subsection{Method \code{get_scientific_object_detail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_object_detail(uri, experiment, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_object_detail_by_experiments"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_object_detail_by_experiments}{}}}
\subsection{Method \code{get_scientific_object_detail_by_experiments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_object_detail_by_experiments(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_object_variables"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_object_variables}{}}}
\subsection{Method \code{get_scientific_object_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_object_variables(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_objects_children"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_objects_children}{}}}
\subsection{Method \code{get_scientific_objects_children()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_objects_children(
  parent,
  experiment,
  rdf_types,
  name,
  factor_levels,
  facility,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_objects_list_by_uris"></a>}}
\if{latex}{\out{\hypertarget{method-get_scientific_objects_list_by_uris}{}}}
\subsection{Method \code{get_scientific_objects_list_by_uris()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_objects_list_by_uris(experiment, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_used_types"></a>}}
\if{latex}{\out{\hypertarget{method-get_used_types}{}}}
\subsection{Method \code{get_used_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_used_types(experiment, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_csv"></a>}}
\if{latex}{\out{\hypertarget{method-import_csv}{}}}
\subsection{Method \code{import_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$import_csv(description, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_scientific_objects"></a>}}
\if{latex}{\out{\hypertarget{method-search_scientific_objects}{}}}
\subsection{Method \code{search_scientific_objects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$search_scientific_objects(
  experiment,
  rdf_types,
  name,
  parent,
  germplasm,
  factor_levels,
  facility,
  existence_date,
  creation_date,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_scientific_objects_with_geometry_list_by_uris"></a>}}
\if{latex}{\out{\hypertarget{method-search_scientific_objects_with_geometry_list_by_uris}{}}}
\subsection{Method \code{search_scientific_objects_with_geometry_list_by_uris()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$search_scientific_objects_with_geometry_list_by_uris(
  experiment,
  start_date,
  end_date,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_scientific_object"></a>}}
\if{latex}{\out{\hypertarget{method-update_scientific_object}{}}}
\subsection{Method \code{update_scientific_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$update_scientific_object(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_csv2"></a>}}
\if{latex}{\out{\hypertarget{method-validate_csv2}{}}}
\subsection{Method \code{validate_csv2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$validate_csv2(description, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
