% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectSearchDTO.r
\docType{data}
\name{ScientificObjectSearchDTO}
\alias{ScientificObjectSearchDTO}
\title{ScientificObjectSearchDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectSearchDTO
}
\description{
ScientificObjectSearchDTO Class
}
\section{Fields}{

\describe{
\item{\code{uris}}{}

\item{\code{excluded_uris}}{}

\item{\code{experiment}}{}

\item{\code{rdf_types}}{}

\item{\code{name}}{}

\item{\code{parent}}{}

\item{\code{germplasm}}{}

\item{\code{factor_levels}}{}

\item{\code{facility}}{}

\item{\code{existence_date}}{}

\item{\code{creation_date}}{}

\item{\code{order_by}}{}

\item{\code{page}}{}

\item{\code{page_sze}}{}
}}

\keyword{datasets}
