% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Season.r
\docType{data}
\name{Season}
\alias{Season}
\title{Season Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Season
}
\description{
Season Class
}
\section{Fields}{

\describe{
\item{\code{season}}{}

\item{\code{seasonDbId}}{}

\item{\code{year}}{}
}}

\keyword{datasets}
