% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecurityApi.r
\docType{data}
\name{SecurityApi}
\alias{SecurityApi}
\title{Security operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SecurityApi
}
\description{
opensilexClientToolsR.Security
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_group Add a group


create_profile Add a profile


create_user Add an user


delete_group Delete a group


delete_profile Delete a profile


delete_user Delete an user


get_all_profiles Get all profiles


get_group Get a group


get_groups_by_uri Get groups by their URIs


get_profile Get a profile


get_user Get an user


get_user_groups Get groups of an user


get_users_by_uri Get users by their URIs


search_groups Search groups


search_profiles Search profiles


search_users Search users


update_group Update a group


update_profile Update a profile


update_user Update an user

}
}

\keyword{datasets}
