% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyDetailsDTO.r
\docType{data}
\name{StudyDetailsDTO}
\alias{StudyDetailsDTO}
\title{StudyDetailsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudyDetailsDTO
}
\description{
StudyDetailsDTO Class
}
\section{Fields}{

\describe{
\item{\code{active}}{}

\item{\code{additionalInfo}}{}

\item{\code{commonCropName}}{}

\item{\code{documentationURL}}{}

\item{\code{endDate}}{}

\item{\code{locationDbId}}{}

\item{\code{locationName}}{}

\item{\code{name}}{}

\item{\code{programDbId}}{}

\item{\code{programName}}{}

\item{\code{seasons}}{}

\item{\code{startDate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{studyType}}{}

\item{\code{studyTypeDbId}}{}

\item{\code{studyTypeName}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}

\item{\code{contacts}}{}

\item{\code{datalinks}}{}

\item{\code{location}}{}
}}

\keyword{datasets}
