% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetPositionGetDTO.r
\docType{data}
\name{TargetPositionGetDTO}
\alias{TargetPositionGetDTO}
\title{TargetPositionGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TargetPositionGetDTO
}
\description{
TargetPositionGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{target}}{}

\item{\code{position}}{}
}}

\keyword{datasets}
