% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trait.r
\docType{data}
\name{Trait}
\alias{Trait}
\title{Trait Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Trait
}
\description{
Trait Class
}
\section{Fields}{

\describe{
\item{\code{alternativeAbbreviations}}{}

\item{\code{attribute}}{}

\item{\code{description}}{}

\item{\code{entity}}{}

\item{\code{mainAbbreviation}}{}

\item{\code{name}}{}

\item{\code{ontologyReference}}{}

\item{\code{status}}{}

\item{\code{synonyms}}{}

\item{\code{traitDbId}}{}

\item{\code{traitName}}{}

\item{\code{xref}}{}

\item{\code{class}}{}
}}

\keyword{datasets}
