% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitCreationDTO.r
\name{UnitCreationDTO}
\alias{UnitCreationDTO}
\title{UnitCreationDTO Class}
\description{
UnitCreationDTO Class

UnitCreationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{symbol}}{}

\item{\code{alternative_symbol}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UnitCreationDTO$new()}}
\item \href{#method-toJSON}{\code{UnitCreationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{UnitCreationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{UnitCreationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{UnitCreationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UnitCreationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{UnitCreationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$new(
  uri,
  name,
  description,
  symbol,
  alternative_symbol,
  exact_match,
  close_match,
  broad_match,
  narrow_match
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$fromJSON(UnitCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$fromJSONObject(UnitCreationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$fromJSONString(UnitCreationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitCreationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
