% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URIsListPostDTO.r
\docType{data}
\name{URIsListPostDTO}
\alias{URIsListPostDTO}
\title{URIsListPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
URIsListPostDTO
}
\description{
URIsListPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{uris}}{}
}}

\keyword{datasets}
