% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URITypesDTO.r
\docType{data}
\name{URITypesDTO}
\alias{URITypesDTO}
\title{URITypesDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
URITypesDTO
}
\description{
URITypesDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdf_types}}{}
}}

\keyword{datasets}
