% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableCreationDTO.r
\docType{data}
\name{VariableCreationDTO}
\alias{VariableCreationDTO}
\title{VariableCreationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariableCreationDTO
}
\description{
VariableCreationDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{alternative_name}}{}

\item{\code{description}}{}

\item{\code{entity}}{}

\item{\code{characteristic}}{}

\item{\code{trait}}{}

\item{\code{trait_name}}{}

\item{\code{method}}{}

\item{\code{unit}}{}

\item{\code{datatype}}{}

\item{\code{time_interval}}{}

\item{\code{sampling_interval}}{}

\item{\code{exact_match}}{}

\item{\code{close_match}}{}

\item{\code{broad_match}}{}

\item{\code{narrow_match}}{}
}}

\keyword{datasets}
