% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesApi.r
\docType{data}
\name{VariablesApi}
\alias{VariablesApi}
\title{Variables operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariablesApi
}
\description{
opensilexClientToolsR.Variables
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_characteristic Add a characteristic


create_entity Add an entity


create_method Add a method


create_unit Add an unit


create_variable Add a variable


delete_characteristic Delete a characteristic


delete_entity Delete an entity


delete_method Delete a method


delete_unit Delete an unit


delete_variable Delete a variable


get_characteristic Get a characteristic


get_datatypes Get variables datatypes


get_entity Get an entity


get_method Get a method


get_unit Get an unit


get_variable Get a variable


get_variables_by_ur_is Get detailed variables by uris


search_characteristics Search characteristics by name


search_entities Search entities by name


search_methods Search methods by name


search_units Search units by name


search_variables Search variables by name, long name, entity, characteristic, method or unit name


search_variables_details Search detailed variables by name, long name, entity, characteristic, method or unit name


update_characteristic Update a characteristic


update_entity Update an entity


update_method Update a method


update_unit Update an unit


update_variable Update a variable

}
}

\keyword{datasets}
