% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesApi.r
\name{VariablesApi}
\alias{VariablesApi}
\title{Variables operations}
\description{
opensilexClientToolsR.Variables
}
\section{Methods}{

\describe{

create_characteristic Add a characteristic


create_entity Add an entity


create_method Add a method


create_unit Add an unit


create_variable Add a variable


create_variables_group Add a variables group


delete_characteristic Delete a characteristic


delete_entity Delete an entity


delete_method Delete a method


delete_unit Delete an unit


delete_variable Delete a variable


delete_variables_group Delete a variables group


get_characteristic Get a characteristic


get_datatypes Get variables datatypes


get_entity Get an entity


get_method Get a method


get_unit Get an unit


get_variable Get a variable


get_variables_by_ur_is Get detailed variables by uris


get_variables_group Get a variables group


get_variables_group_by_ur_is Get variables groups by their URIs


search_characteristics Search characteristics by name


search_entities Search entities by name


search_methods Search methods by name


search_units Search units by name


search_variables Search variables by name, long name, entity, characteristic, method or unit name


search_variables_details Search detailed variables by name, long name, entity, characteristic, method or unit name


search_variables_groups Search variables groups


update_characteristic Update a characteristic


update_entity Update an entity


update_method Update a method


update_unit Update an unit


update_variable Update a variable


update_variables_group Update a variables group

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariablesApi$new()}}
\item \href{#method-create_characteristic}{\code{VariablesApi$create_characteristic()}}
\item \href{#method-create_entity}{\code{VariablesApi$create_entity()}}
\item \href{#method-create_method}{\code{VariablesApi$create_method()}}
\item \href{#method-create_unit}{\code{VariablesApi$create_unit()}}
\item \href{#method-create_variable}{\code{VariablesApi$create_variable()}}
\item \href{#method-create_variables_group}{\code{VariablesApi$create_variables_group()}}
\item \href{#method-delete_characteristic}{\code{VariablesApi$delete_characteristic()}}
\item \href{#method-delete_entity}{\code{VariablesApi$delete_entity()}}
\item \href{#method-delete_method}{\code{VariablesApi$delete_method()}}
\item \href{#method-delete_unit}{\code{VariablesApi$delete_unit()}}
\item \href{#method-delete_variable}{\code{VariablesApi$delete_variable()}}
\item \href{#method-delete_variables_group}{\code{VariablesApi$delete_variables_group()}}
\item \href{#method-get_characteristic}{\code{VariablesApi$get_characteristic()}}
\item \href{#method-get_datatypes}{\code{VariablesApi$get_datatypes()}}
\item \href{#method-get_entity}{\code{VariablesApi$get_entity()}}
\item \href{#method-get_method}{\code{VariablesApi$get_method()}}
\item \href{#method-get_unit}{\code{VariablesApi$get_unit()}}
\item \href{#method-get_variable}{\code{VariablesApi$get_variable()}}
\item \href{#method-get_variables_by_ur_is}{\code{VariablesApi$get_variables_by_ur_is()}}
\item \href{#method-get_variables_group}{\code{VariablesApi$get_variables_group()}}
\item \href{#method-get_variables_group_by_ur_is}{\code{VariablesApi$get_variables_group_by_ur_is()}}
\item \href{#method-search_characteristics}{\code{VariablesApi$search_characteristics()}}
\item \href{#method-search_entities}{\code{VariablesApi$search_entities()}}
\item \href{#method-search_methods}{\code{VariablesApi$search_methods()}}
\item \href{#method-search_units}{\code{VariablesApi$search_units()}}
\item \href{#method-search_variables}{\code{VariablesApi$search_variables()}}
\item \href{#method-search_variables_details}{\code{VariablesApi$search_variables_details()}}
\item \href{#method-search_variables_groups}{\code{VariablesApi$search_variables_groups()}}
\item \href{#method-update_characteristic}{\code{VariablesApi$update_characteristic()}}
\item \href{#method-update_entity}{\code{VariablesApi$update_entity()}}
\item \href{#method-update_method}{\code{VariablesApi$update_method()}}
\item \href{#method-update_unit}{\code{VariablesApi$update_unit()}}
\item \href{#method-update_variable}{\code{VariablesApi$update_variable()}}
\item \href{#method-update_variables_group}{\code{VariablesApi$update_variables_group()}}
\item \href{#method-clone}{\code{VariablesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_characteristic"></a>}}
\if{latex}{\out{\hypertarget{method-create_characteristic}{}}}
\subsection{Method \code{create_characteristic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$create_characteristic(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_entity"></a>}}
\if{latex}{\out{\hypertarget{method-create_entity}{}}}
\subsection{Method \code{create_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$create_entity(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_method"></a>}}
\if{latex}{\out{\hypertarget{method-create_method}{}}}
\subsection{Method \code{create_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$create_method(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_unit"></a>}}
\if{latex}{\out{\hypertarget{method-create_unit}{}}}
\subsection{Method \code{create_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$create_unit(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_variable"></a>}}
\if{latex}{\out{\hypertarget{method-create_variable}{}}}
\subsection{Method \code{create_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$create_variable(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_variables_group"></a>}}
\if{latex}{\out{\hypertarget{method-create_variables_group}{}}}
\subsection{Method \code{create_variables_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$create_variables_group(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_characteristic"></a>}}
\if{latex}{\out{\hypertarget{method-delete_characteristic}{}}}
\subsection{Method \code{delete_characteristic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$delete_characteristic(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_entity"></a>}}
\if{latex}{\out{\hypertarget{method-delete_entity}{}}}
\subsection{Method \code{delete_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$delete_entity(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_method"></a>}}
\if{latex}{\out{\hypertarget{method-delete_method}{}}}
\subsection{Method \code{delete_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$delete_method(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_unit"></a>}}
\if{latex}{\out{\hypertarget{method-delete_unit}{}}}
\subsection{Method \code{delete_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$delete_unit(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_variable"></a>}}
\if{latex}{\out{\hypertarget{method-delete_variable}{}}}
\subsection{Method \code{delete_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$delete_variable(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_variables_group"></a>}}
\if{latex}{\out{\hypertarget{method-delete_variables_group}{}}}
\subsection{Method \code{delete_variables_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$delete_variables_group(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_characteristic"></a>}}
\if{latex}{\out{\hypertarget{method-get_characteristic}{}}}
\subsection{Method \code{get_characteristic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_characteristic(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datatypes"></a>}}
\if{latex}{\out{\hypertarget{method-get_datatypes}{}}}
\subsection{Method \code{get_datatypes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_datatypes(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_entity"></a>}}
\if{latex}{\out{\hypertarget{method-get_entity}{}}}
\subsection{Method \code{get_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_entity(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_method"></a>}}
\if{latex}{\out{\hypertarget{method-get_method}{}}}
\subsection{Method \code{get_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_method(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_unit"></a>}}
\if{latex}{\out{\hypertarget{method-get_unit}{}}}
\subsection{Method \code{get_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_unit(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable"></a>}}
\if{latex}{\out{\hypertarget{method-get_variable}{}}}
\subsection{Method \code{get_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variable(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_by_ur_is"></a>}}
\if{latex}{\out{\hypertarget{method-get_variables_by_ur_is}{}}}
\subsection{Method \code{get_variables_by_ur_is()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variables_by_ur_is(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_group"></a>}}
\if{latex}{\out{\hypertarget{method-get_variables_group}{}}}
\subsection{Method \code{get_variables_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variables_group(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_group_by_ur_is"></a>}}
\if{latex}{\out{\hypertarget{method-get_variables_group_by_ur_is}{}}}
\subsection{Method \code{get_variables_group_by_ur_is()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variables_group_by_ur_is(uris, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-search_characteristics}{}}}
\subsection{Method \code{search_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_characteristics(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_entities"></a>}}
\if{latex}{\out{\hypertarget{method-search_entities}{}}}
\subsection{Method \code{search_entities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_entities(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_methods"></a>}}
\if{latex}{\out{\hypertarget{method-search_methods}{}}}
\subsection{Method \code{search_methods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_methods(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_units"></a>}}
\if{latex}{\out{\hypertarget{method-search_units}{}}}
\subsection{Method \code{search_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_units(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_variables"></a>}}
\if{latex}{\out{\hypertarget{method-search_variables}{}}}
\subsection{Method \code{search_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_variables(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_variables_details"></a>}}
\if{latex}{\out{\hypertarget{method-search_variables_details}{}}}
\subsection{Method \code{search_variables_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_variables_details(name, order_by, page, page_size, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_variables_groups"></a>}}
\if{latex}{\out{\hypertarget{method-search_variables_groups}{}}}
\subsection{Method \code{search_variables_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$search_variables_groups(
  name,
  variable_uri,
  order_by,
  page,
  page_size,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_characteristic"></a>}}
\if{latex}{\out{\hypertarget{method-update_characteristic}{}}}
\subsection{Method \code{update_characteristic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$update_characteristic(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_entity"></a>}}
\if{latex}{\out{\hypertarget{method-update_entity}{}}}
\subsection{Method \code{update_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$update_entity(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_method"></a>}}
\if{latex}{\out{\hypertarget{method-update_method}{}}}
\subsection{Method \code{update_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$update_method(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_unit"></a>}}
\if{latex}{\out{\hypertarget{method-update_unit}{}}}
\subsection{Method \code{update_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$update_unit(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_variable"></a>}}
\if{latex}{\out{\hypertarget{method-update_variable}{}}}
\subsection{Method \code{update_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$update_variable(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_variables_group"></a>}}
\if{latex}{\out{\hypertarget{method-update_variables_group}{}}}
\subsection{Method \code{update_variables_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$update_variables_group(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
