% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionInfoDTO.r
\docType{data}
\name{VersionInfoDTO}
\alias{VersionInfoDTO}
\title{VersionInfoDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VersionInfoDTO
}
\description{
VersionInfoDTO Class
}
\section{Fields}{

\describe{
\item{\code{title}}{}

\item{\code{version}}{}

\item{\code{description}}{}

\item{\code{contact}}{}

\item{\code{license}}{}

\item{\code{modules_version}}{}

\item{\code{external_docs}}{}

\item{\code{api_docs}}{}

\item{\code{git_commit}}{}

\item{\code{github_page}}{}
}}

\keyword{datasets}
