% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueDataTypeDTO.r
\docType{data}
\name{VueDataTypeDTO}
\alias{VueDataTypeDTO}
\title{VueDataTypeDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VueDataTypeDTO
}
\description{
VueDataTypeDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{short_uri}}{}

\item{\code{input_component}}{}

\item{\code{view_component}}{}

\item{\code{label_key}}{}
}}

\keyword{datasets}
