% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueJsApi.r
\name{VueJsApi}
\alias{VueJsApi}
\title{VueJs operations}
\description{
opensilexClientToolsR.VueJs
}
\section{Methods}{

\describe{

get_config Return the current configuration


get_extension Return the front Vue JS extension file to include


get_extension_style Return the front Vue JS extension css file to include


get_theme_config Return the front Vue JS theme configuration


get_theme_css Return the theme css file


get_theme_resource Return the theme requested resource

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueJsApi$new()}}
\item \href{#method-get_config}{\code{VueJsApi$get_config()}}
\item \href{#method-get_extension}{\code{VueJsApi$get_extension()}}
\item \href{#method-get_extension_style}{\code{VueJsApi$get_extension_style()}}
\item \href{#method-get_theme_config}{\code{VueJsApi$get_theme_config()}}
\item \href{#method-get_theme_css}{\code{VueJsApi$get_theme_css()}}
\item \href{#method-get_theme_resource}{\code{VueJsApi$get_theme_resource()}}
\item \href{#method-clone}{\code{VueJsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$get_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_extension"></a>}}
\if{latex}{\out{\hypertarget{method-get_extension}{}}}
\subsection{Method \code{get_extension()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$get_extension(module, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_extension_style"></a>}}
\if{latex}{\out{\hypertarget{method-get_extension_style}{}}}
\subsection{Method \code{get_extension_style()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$get_extension_style(module, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme_config}{}}}
\subsection{Method \code{get_theme_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$get_theme_config(module_id, theme_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme_css"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme_css}{}}}
\subsection{Method \code{get_theme_css()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$get_theme_css(module_id, theme_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theme_resource"></a>}}
\if{latex}{\out{\hypertarget{method-get_theme_resource}{}}}
\subsection{Method \code{get_theme_resource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$get_theme_resource(module_id, theme_id, file_path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
