% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueJsOntologyExtensionApi.r
\name{VueJsOntologyExtensionApi}
\alias{VueJsOntologyExtensionApi}
\title{VueJsOntologyExtension operations}
\description{
opensilexClientToolsR.VueJsOntologyExtension
}
\section{Methods}{

\describe{

create_rdf_type Create a custom class


delete_rdf_type Delete a RDF type


get_data_types1 Return literal datatypes definition


get_object_types Return object types definition


get_rdf_type1 Return rdf type model definition with properties


get_rdf_type_properties Return class model properties definitions


get_rdf_types_parameters Return RDF types parameters for Vue.js application


set_rdf_type_properties_order Define properties order


update_rdf_type Update a custom class

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueJsOntologyExtensionApi$new()}}
\item \href{#method-create_rdf_type}{\code{VueJsOntologyExtensionApi$create_rdf_type()}}
\item \href{#method-delete_rdf_type}{\code{VueJsOntologyExtensionApi$delete_rdf_type()}}
\item \href{#method-get_data_types1}{\code{VueJsOntologyExtensionApi$get_data_types1()}}
\item \href{#method-get_object_types}{\code{VueJsOntologyExtensionApi$get_object_types()}}
\item \href{#method-get_rdf_type1}{\code{VueJsOntologyExtensionApi$get_rdf_type1()}}
\item \href{#method-get_rdf_type_properties}{\code{VueJsOntologyExtensionApi$get_rdf_type_properties()}}
\item \href{#method-get_rdf_types_parameters}{\code{VueJsOntologyExtensionApi$get_rdf_types_parameters()}}
\item \href{#method-set_rdf_type_properties_order}{\code{VueJsOntologyExtensionApi$set_rdf_type_properties_order()}}
\item \href{#method-update_rdf_type}{\code{VueJsOntologyExtensionApi$update_rdf_type()}}
\item \href{#method-clone}{\code{VueJsOntologyExtensionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_rdf_type"></a>}}
\if{latex}{\out{\hypertarget{method-create_rdf_type}{}}}
\subsection{Method \code{create_rdf_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$create_rdf_type(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_rdf_type"></a>}}
\if{latex}{\out{\hypertarget{method-delete_rdf_type}{}}}
\subsection{Method \code{delete_rdf_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$delete_rdf_type(rdf_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_types1"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_types1}{}}}
\subsection{Method \code{get_data_types1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$get_data_types1(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_object_types"></a>}}
\if{latex}{\out{\hypertarget{method-get_object_types}{}}}
\subsection{Method \code{get_object_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$get_object_types(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rdf_type1"></a>}}
\if{latex}{\out{\hypertarget{method-get_rdf_type1}{}}}
\subsection{Method \code{get_rdf_type1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$get_rdf_type1(rdf_type, parent_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rdf_type_properties"></a>}}
\if{latex}{\out{\hypertarget{method-get_rdf_type_properties}{}}}
\subsection{Method \code{get_rdf_type_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$get_rdf_type_properties(rdf_type, parent_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rdf_types_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-get_rdf_types_parameters}{}}}
\subsection{Method \code{get_rdf_types_parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$get_rdf_types_parameters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rdf_type_properties_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_rdf_type_properties_order}{}}}
\subsection{Method \code{set_rdf_type_properties_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$set_rdf_type_properties_order(rdf_type, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_rdf_type"></a>}}
\if{latex}{\out{\hypertarget{method-update_rdf_type}{}}}
\subsection{Method \code{update_rdf_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$update_rdf_type(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueJsOntologyExtensionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
