% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueObjectTypeDTO.r
\name{VueObjectTypeDTO}
\alias{VueObjectTypeDTO}
\title{VueObjectTypeDTO Class}
\description{
VueObjectTypeDTO Class

VueObjectTypeDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{short_uri}}{}

\item{\code{input_component}}{}

\item{\code{input_components_by_property}}{}

\item{\code{view_component}}{}

\item{\code{rdf_type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueObjectTypeDTO$new()}}
\item \href{#method-toJSON}{\code{VueObjectTypeDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VueObjectTypeDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VueObjectTypeDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VueObjectTypeDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VueObjectTypeDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VueObjectTypeDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$new(
  uri,
  name,
  short_uri,
  input_component,
  input_components_by_property,
  view_component,
  rdf_type
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$fromJSON(VueObjectTypeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$fromJSONObject(VueObjectTypeDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$fromJSONString(VueObjectTypeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueObjectTypeDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
