% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueRDFTypeDTO.r
\name{VueRDFTypeDTO}
\alias{VueRDFTypeDTO}
\title{VueRDFTypeDTO Class}
\description{
VueRDFTypeDTO Class

VueRDFTypeDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{parent}}{}

\item{\code{icon}}{}

\item{\code{name_translations}}{}

\item{\code{comment_translations}}{}

\item{\code{is_abstract}}{}

\item{\code{data_properties}}{}

\item{\code{object_properties}}{}

\item{\code{properties_order}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueRDFTypeDTO$new()}}
\item \href{#method-toJSON}{\code{VueRDFTypeDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VueRDFTypeDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VueRDFTypeDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VueRDFTypeDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VueRDFTypeDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VueRDFTypeDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$new(
  uri,
  name,
  comment,
  parent,
  icon,
  name_translations,
  comment_translations,
  is_abstract,
  data_properties,
  object_properties,
  properties_order
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$fromJSON(VueRDFTypeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$fromJSONObject(VueRDFTypeDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$fromJSONString(VueRDFTypeDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
