% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueRDFTypeParameterDTO.r
\name{VueRDFTypeParameterDTO}
\alias{VueRDFTypeParameterDTO}
\title{VueRDFTypeParameterDTO Class}
\description{
VueRDFTypeParameterDTO Class

VueRDFTypeParameterDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{icon}}{}

\item{\code{is_abstract}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueRDFTypeParameterDTO$new()}}
\item \href{#method-toJSON}{\code{VueRDFTypeParameterDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VueRDFTypeParameterDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VueRDFTypeParameterDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VueRDFTypeParameterDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VueRDFTypeParameterDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VueRDFTypeParameterDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$new(uri, icon, is_abstract)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$fromJSON(VueRDFTypeParameterDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$fromJSONObject(VueRDFTypeParameterDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$fromJSONString(VueRDFTypeParameterDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypeParameterDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
