% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VueRDFTypePropertyDTO.r
\name{VueRDFTypePropertyDTO}
\alias{VueRDFTypePropertyDTO}
\title{VueRDFTypePropertyDTO Class}
\description{
VueRDFTypePropertyDTO Class

VueRDFTypePropertyDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{property}}{}

\item{\code{name}}{}

\item{\code{comment}}{}

\item{\code{inherited}}{}

\item{\code{target_property}}{}

\item{\code{input_component}}{}

\item{\code{input_components_by_property}}{}

\item{\code{view_component}}{}

\item{\code{is_list}}{}

\item{\code{is_required}}{}

\item{\code{is_custom}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VueRDFTypePropertyDTO$new()}}
\item \href{#method-toJSON}{\code{VueRDFTypePropertyDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VueRDFTypePropertyDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VueRDFTypePropertyDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VueRDFTypePropertyDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VueRDFTypePropertyDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VueRDFTypePropertyDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$new(
  property,
  name,
  comment,
  inherited,
  target_property,
  input_component,
  input_components_by_property,
  view_component,
  is_list,
  is_required,
  is_custom
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$fromJSON(VueRDFTypePropertyDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$fromJSONObject(VueRDFTypePropertyDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$fromJSONString(VueRDFTypePropertyDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VueRDFTypePropertyDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
